refactor: WorldModel Phases 2+3 — extract reasoning, introspection, context
Phase 2 — Philosophical + Creative + General reasoning (9 modules):
- philosophical_reasoning.py, ethical_dilemma.py, ethical_principles.py
- ethical_conflicts.py, ethical_synthesis.py, ethical_values.py
- creative_reasoning.py, general_reasoning.py, general_delegation.py
- 1,613 lines extracted, 30 methods delegated

Phase 3 — Introspection + Context + Routing + Performance (9 modules):
- introspection_core.py, introspection_self.py, introspection_meta.py
- introspection_analysis.py, introspection_domain.py, introspection_demo.py
- context_enrich.py, routing_recommend.py, performance_introspect.py
- 1,927 lines extracted, 40 methods delegated

world_model_core.py: 7,339 → 3,799 lines (-3,540, -48%).
Phases 2+3 of 8 in deep God class decomposition.
