+++ src/vulcan/world_model/ethical_conflicts.py
@@ -1,0 +1,242 @@
+"""
+Ethical conflict detection and counterfactual analysis functions extracted from WorldModel.
+
+Handles detecting principle conflicts, running ethical boundary analysis,
+detecting goal conflicts, and analyzing option counterfactuals.
+"""
+
+import logging
+from typing import Any, Dict, List, Optional, Tuple
+
+logger = logging.getLogger(__name__)
+
+
+def _detect_principle_conflicts(
+    wm,
+    option_analysis: List[Dict[str, Any]],
+    principles: List[Dict[str, Any]]
+) -> List[Dict[str, Any]]:
+    """
+    Detect conflicts between moral principles based on option analysis.
+
+    Industry Standard: Identify genuine ethical dilemmas where principles clash.
+
+    Args:
+        option_analysis: Analysis of options from analyze_options_against_principles
+        principles: List of principles from extract_moral_principles
+
+    Returns:
+        List of conflict dictionaries with descriptions
+    """
+    try:
+        conflicts = []
+
+        # Check if different principles are violated by different options
+        principle_names = [p['name'] for p in principles]
+
+        # Conflict detection: If no option satisfies all principles
+        if len(option_analysis) >= 2:
+            # Check if option A violates different principles than option B
+            option_a = option_analysis[0]
+            option_b = option_analysis[1] if len(option_analysis) > 1 else {}
+
+            a_violations = set(v.split('by')[0].strip() for v in option_a.get('violations', []))
+            b_violations = set(v.split('by')[0].strip() for v in option_b.get('violations', []))
+
+            if a_violations != b_violations and (a_violations or b_violations):
+                conflicts.append({
+                    'description': 'Different options violate different principles - genuine dilemma',
+                    'type': 'principle_tradeoff'
+                })
+
+        # Specific conflicts
+        if 'non-instrumentalization' in principle_names and 'non-negligence' in principle_names:
+            conflicts.append({
+                'description': 'Non-instrumentalization conflicts with non-negligence when intervention saves more lives',
+                'type': 'action_vs_inaction'
+            })
+
+        if 'non-maleficence' in principle_names and 'beneficence' in principle_names:
+            conflicts.append({
+                'description': 'Preventing harm (beneficence) may require causing lesser harm (maleficence)',
+                'type': 'harm_vs_benefit'
+            })
+
+        return conflicts
+
+    except Exception as e:
+        logger.warning(f"[WorldModel] Conflict detection error: {e}")
+        return []
+
+
+def _run_ethical_boundary_analysis(
+    wm, structure: Dict[str, Any], query: str
+) -> Dict[str, Any]:
+    """Run EthicalBoundaryMonitor on each option."""
+    if not wm.ethical_boundary_monitor:
+        return {'analysis': 'not_available'}
+
+    results = {}
+    for option in structure.get('options', ['general_action']):
+        try:
+            # Check ethical boundaries for this option
+            boundary_check = wm.ethical_boundary_monitor.check_action(
+                action={'type': option, 'query': query},
+                context={'ethical_structure': structure}
+            )
+            results[option] = boundary_check
+        except Exception as e:
+            logger.debug(f"Ethical boundary check failed for {option}: {e}")
+            results[option] = {'status': 'check_failed'}
+
+    return results
+
+
+def _detect_goal_conflicts_in_query(
+    wm, structure: Dict[str, Any], query: str
+) -> Dict[str, Any]:
+    """Detect goal conflicts using GoalConflictDetector."""
+    if not wm.goal_conflict_detector:
+        return {'conflicts': 'not_detected'}
+
+    try:
+        # Extract goals from options
+        goals = []
+        if 'trolley' in structure.get('type', ''):
+            goals = [
+                {'name': 'minimize_deaths', 'priority': 'high'},
+                {'name': 'avoid_direct_harm', 'priority': 'high'}
+            ]
+
+        if goals:
+            conflicts = wm.goal_conflict_detector.detect_conflicts(
+                goals, context={'query': query}
+            )
+            return conflicts
+    except Exception as e:
+        logger.debug(f"Goal conflict detection failed: {e}")
+
+    return {'conflicts': []}
+
+
+def _analyze_option_counterfactuals(
+    wm, structure: Dict[str, Any]
+) -> Optional[Dict[str, Any]]:
+    """Analyze counterfactual outcomes for each option."""
+    if not wm.counterfactual_reasoner:
+        return None
+
+    try:
+        results = {}
+        for option in structure.get('options', []):
+            outcome = wm.counterfactual_reasoner.predict_under_objective(
+                alternative_objective=option,
+                context={'ethical_structure': structure}
+            )
+            if outcome:
+                results[option] = {
+                    'predicted_value': outcome.predicted_value if hasattr(outcome, 'predicted_value') else None,
+                    'confidence': outcome.confidence if hasattr(outcome, 'confidence') else 0.5,
+                    'side_effects': outcome.side_effects if hasattr(outcome, 'side_effects') else {}
+                }
+        return results if results else None
+    except Exception as e:
+        logger.debug(f"Counterfactual analysis failed: {e}")
+        return None
+
+
+def _synthesize_dilemma_decision(
+    wm,
+    structure: Dict[str, Any],
+    principles: List[Dict[str, Any]],
+    option_analysis: List[Dict[str, Any]],
+    conflicts: List[Dict[str, Any]],
+    query: str
+) -> Tuple[str, str]:
+    """
+    Synthesize a reasoned decision from the analysis.
+
+    Industry Standard: Multi-factor decision making with clear rationale.
+
+    Args:
+        structure: Parsed dilemma structure
+        principles: Extracted moral principles
+        option_analysis: Analysis of each option
+        conflicts: Detected principle conflicts
+        query: Original query
+
+    Returns:
+        Tuple of (decision string, reasoning string)
+    """
+    try:
+        query_lower = query.lower()
+
+        # Decision logic based on utilitarian + deontological balance
+        # Industry Standard: Prefer option that minimizes total harm
+
+        # Count violations and compliances for each option
+        option_scores = []
+        for analysis in option_analysis:
+            score = len(analysis.get('compliances', [])) - len(analysis.get('violations', []))
+            option_scores.append({
+                'option': analysis['option'],
+                'score': score,
+                'description': analysis['description']
+            })
+
+        # Choose option with best score
+        if option_scores:
+            best_option = max(option_scores, key=lambda x: x['score'])
+            decision = f"{best_option['option']}. {best_option['description']}"
+        else:
+            # Fallback: Choose A by default
+            decision = "A. Pull the lever"
+
+        # Build reasoning
+        reasoning_parts = []
+
+        # Principle analysis
+        if principles:
+            reasoning_parts.append("**Moral Principles Considered:**")
+            for p in principles[:3]:  # Top 3
+                reasoning_parts.append(f"- {p['name']}: {p['description']}")
+
+        # Option analysis
+        reasoning_parts.append("\n**Analysis of Options:**")
+        for analysis in option_analysis:
+            option_id = analysis['option']
+            reasoning_parts.append(f"\n*Option {option_id}:*")
+            if analysis.get('compliances'):
+                reasoning_parts.append(f"- Compliance: {'; '.join(analysis['compliances'][:2])}")
+            if analysis.get('violations'):
+                reasoning_parts.append(f"- Conflicts: {'; '.join(analysis['violations'][:2])}")
+
+        # Conflict resolution
+        if conflicts:
+            reasoning_parts.append("\n**Conflict Resolution:**")
+            for conflict in conflicts:
+                reasoning_parts.append(f"- {conflict['description']}")
+
+        # Consequentialist analysis (numbers)
+        if 'five' in query_lower and 'one' in query_lower:
+            reasoning_parts.append("\n**Consequentialist Perspective:**")
+            reasoning_parts.append("- From a utilitarian view: 1 death < 5 deaths")
+            reasoning_parts.append("- Minimizing total harm favors action to save the greater number")
+
+        # Final justification
+        reasoning_parts.append("\n**Conclusion:**")
+        if 'pull' in decision.lower():
+            reasoning_parts.append("- The principle of non-negligence (duty to prevent harm) takes priority")
+            reasoning_parts.append("- While pulling the lever involves direct action causing one death,")
+            reasoning_parts.append("- Inaction allowing five deaths is morally worse when prevention is possible")
+        else:
+            reasoning_parts.append("- The principle of non-instrumentalization takes priority")
+            reasoning_parts.append("- Using one person as a means to save others violates their dignity")
+
+        reasoning = "\n".join(reasoning_parts)
+
+        return decision, reasoning
+
+    except Exception as e:
+        logger.warning(f"[WorldModel] Decision synthesis error: {e}")
+        return ("Unable to reach decision", "Analysis incomplete due to error")
