+++ src/vulcan/world_model/world_model_core.py
@@ -2600,310 +2600,59 @@ class WorldModel:
     PATH_STRENGTH_THRESHOLD = 0.7  # Minimum causal path strength to predict failure
     
     def has_sufficient_history(self, min_observations: int = MIN_OBSERVATIONS_FOR_RECOMMENDATIONS) -> bool:
-        """
-        Check if world model has enough data to make recommendations.
-        
-        The world model needs observations before it can provide meaningful
-        routing recommendations based on learned patterns.
-        
-        Args:
-            min_observations: Minimum number of observations needed 
-                              (default MIN_OBSERVATIONS_FOR_RECOMMENDATIONS = 5)
-            
-        Returns:
-            True if sufficient history exists for recommendations
-        """
-        return self.observation_count >= min_observations
-    
-    def recommend_routing(
-        self, 
-        query: str, 
-        classification: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        """
-        Recommend routing based on learned patterns.
-        
-        Consults the world model's causal understanding and observation history
-        to provide routing recommendations. This allows the world model to
-        learn from experience and suggest better tool selections.
-        
-        Args:
-            query: Query text
-            classification: Initial classification from query router
-            
-        Returns:
-            Dict with:
-            - tools: Recommended tools
-            - confidence: Confidence in recommendation  
-            - reasoning: Why these tools recommended
-            - warnings: Potential issues
-            - alternative_routing: Alternative if issues expected
-        """
-        # Check if we have enough history
-        if not self.has_sufficient_history():
-            return {
-                'tools': classification.get('tools', []),
-                'confidence': classification.get('confidence', 0.5),
-                'reasoning': 'Insufficient historical data for recommendations',
-                'warnings': [],
-                'alternative_routing': None
-            }
-        
-        # Extract query features
-        query_features = self._extract_query_features(query)
-        
-        # Check for predicted failures
-        tools = classification.get('tools', [])
-        if self._predicts_failure(query_features, tools):
-            warnings = self._get_failure_warnings(query_features, tools)
-            alternatives = self._suggest_alternative_tools(query_features)
-            
-            return {
-                'tools': alternatives.get('tools', tools),
-                'confidence': alternatives.get('confidence', 0.6),
-                'reasoning': f"Similar queries failed with {tools}. Recommending alternatives.",
-                'warnings': warnings,
-                'alternative_routing': alternatives
-            }
-        
-        # Return classification with world model confidence
-        return {
-            'tools': tools,
-            'confidence': classification.get('confidence', 0.7),
-            'reasoning': 'Classification confirmed by world model patterns',
-            'warnings': [],
-            'alternative_routing': None
-        }
-    
+        """Check if world model has enough data to make recommendations."""
+        from .routing_recommend import has_sufficient_history as _fn
+        return _fn(self, min_observations)
+
+    def recommend_routing(self, query: str, classification: Dict[str, Any]) -> Dict[str, Any]:
+        """Recommend routing based on learned patterns."""
+        from .routing_recommend import recommend_routing as _fn
+        return _fn(self, query, classification)
+
     def _extract_query_features(self, query: str) -> Dict[str, Any]:
         """Extract features from query for pattern matching."""
-        if not query:
-            return {'type': 'unknown'}
-        
-        query_lower = query.lower()
-        
-        return {
-            'type': self._infer_query_type(query_lower),
-            'has_formal_logic': any(sym in query for sym in FORMAL_LOGIC_SYMBOLS),
-            'has_probability': any(kw in query_lower for kw in PROBABILITY_KEYWORDS),
-            'is_self_referential': any(kw in query_lower for kw in SELF_REFERENTIAL_KEYWORDS),
-            'complexity': len(query) / 100.0  # Simple complexity proxy
-        }
-    
+        from .routing_recommend import extract_query_features as _fn
+        return _fn(self, query)
+
     def _infer_query_type(self, query_lower: str) -> str:
         """Infer query type from content using shared constants."""
-        if any(sym in query_lower for sym in FORMAL_LOGIC_SYMBOLS) or any(kw in query_lower for kw in FORMAL_LOGIC_KEYWORDS):
-            return 'formal_logic'
-        elif any(kw in query_lower for kw in PROBABILITY_KEYWORDS):
-            return 'probabilistic'
-        elif any(kw in query_lower for kw in ['compute', 'calculate', 'integral']):
-            return 'mathematical'
-        elif any(kw in query_lower for kw in SELF_REFERENTIAL_KEYWORDS):
-            return 'self_referential'
-        elif any(kw in query_lower for kw in ['cause', 'effect', 'intervention']):
-            return 'causal'
-        else:
-            return 'general'
-    
+        from .routing_recommend import infer_query_type as _fn
+        return _fn(self, query_lower)
+
     def _predicts_failure(self, query_features: Dict[str, Any], tools: List[str]) -> bool:
         """Predict if routing is likely to fail based on causal patterns."""
-        if not CAUSAL_GRAPH_AVAILABLE or not self.causal_graph:
-            return False
-        
-        query_type = query_features.get('type', 'unknown')
-        
-        # Check causal graph for failure patterns
-        for tool in tools:
-            pattern_node = f'{tool}_on_{query_type}'
-            try:
-                if self.causal_graph.has_node(pattern_node):
-                    # Check if there's a strong path to validation_failure
-                    paths = self.causal_graph.find_all_paths(pattern_node, 'validation_failure')
-                    if paths:
-                        for path in paths:
-                            if hasattr(path, 'total_strength') and path.total_strength > self.PATH_STRENGTH_THRESHOLD:
-                                return True
-            except Exception:
-                pass
-        
-        return False
-    
+        from .routing_recommend import predicts_failure as _fn
+        return _fn(self, query_features, tools)
+
     def _get_failure_warnings(self, query_features: Dict[str, Any], tools: List[str]) -> List[str]:
         """Get warnings about potential failures."""
-        warnings = []
-        query_type = query_features.get('type', 'unknown')
-        
-        for tool in tools:
-            # Check for mismatches
-            if query_type == 'formal_logic' and 'math' in tool.lower():
-                warnings.append(f"{tool} may produce incorrect results for formal logic queries")
-            if query_type == 'probabilistic' and 'symbolic' in tool.lower():
-                warnings.append(f"{tool} may not handle probabilistic reasoning well")
-        
-        return warnings
-    
+        from .routing_recommend import get_failure_warnings as _fn
+        return _fn(self, query_features, tools)
+
     def _suggest_alternative_tools(self, query_features: Dict[str, Any]) -> Dict[str, Any]:
         """Suggest alternative tools based on query features."""
-        query_type = query_features.get('type', 'unknown')
-        
-        # Default tool suggestions by query type
-        suggestions = {
-            'formal_logic': {'tools': ['logic_engine', 'symbolic'], 'confidence': 0.8},
-            'probabilistic': {'tools': ['probabilistic', 'bayesian'], 'confidence': 0.8},
-            'mathematical': {'tools': ['mathematical', 'computational'], 'confidence': 0.8},
-            'causal': {'tools': ['causal', 'counterfactual'], 'confidence': 0.8},
-            'self_referential': {'tools': ['world_model', 'meta_reasoning'], 'confidence': 0.9},
-        }
-        
-        return suggestions.get(query_type, {'tools': ['hybrid'], 'confidence': 0.6})
+        from .routing_recommend import suggest_alternative_tools as _fn
+        return _fn(self, query_features)
     
     def introspect_performance(self) -> Dict[str, Any]:
-        """
-        Provide self-knowledge about system performance.
-        
-        This allows the system to answer questions like:
-        - "What's your current accuracy?"
-        - "Which engines are working well?"
-        - "What issues have you encountered?"
-        
-        Returns:
-            Dict with performance metrics and known issues
-        """
-        if self.observation_count < MIN_OBSERVATIONS_FOR_RECOMMENDATIONS:
-            return {
-                'status': 'insufficient_data',
-                'message': 'Need more observations for meaningful performance analysis',
-                'observation_count': self.observation_count
-            }
-        
-        # Compute performance stats
-        stats = self._compute_performance_stats()
-        
-        # Identify known issues
-        issues = self._identify_known_issues()
-        
-        # Assess capabilities
-        capabilities = self._assess_engine_capabilities()
-        
-        return {
-            'status': 'operational',
-            'performance': stats,
-            'known_issues': issues,
-            'capabilities': capabilities,
-            'confidence': 0.95,
-            'observation_count': self.observation_count,
-            'model_version': self.model_version
-        }
-    
+        """Provide self-knowledge about system performance."""
+        from .performance_introspect import introspect_performance as _fn
+        return _fn(self)
+
     def _compute_performance_stats(self) -> Dict[str, Any]:
         """Compute performance statistics from observation history."""
-        # Get stats from observation processor
-        history_size = len(self.observation_processor.observation_history)
-        
-        if history_size == 0:
-            return {'message': 'No observations recorded yet'}
-        
-        # Calculate basic metrics
-        stats = {
-            'total_observations': history_size,
-            'model_version': self.model_version,
-        }
-        
-        # Add prediction history stats if available
-        if hasattr(self, 'prediction_manager') and self.prediction_manager:
-            pred_history = list(self.prediction_manager.prediction_history)
-            if pred_history:
-                confidences = [
-                    p['prediction'].confidence 
-                    for p in pred_history 
-                    if hasattr(p.get('prediction'), 'confidence')
-                ]
-                if confidences:
-                    stats['avg_prediction_confidence'] = sum(confidences) / len(confidences)
-                    stats['prediction_count'] = len(pred_history)
-        
-        # Add causal graph stats
-        if CAUSAL_GRAPH_AVAILABLE and self.causal_graph:
-            stats['causal_nodes'] = len(self.causal_graph.nodes) if hasattr(self.causal_graph, 'nodes') else 0
-            stats['causal_edges'] = len(self.causal_graph.edges) if hasattr(self.causal_graph, 'edges') else 0
-        
-        return stats
-    
+        from .performance_introspect import compute_performance_stats as _fn
+        return _fn(self)
+
     def _identify_known_issues(self) -> List[Dict[str, Any]]:
         """Identify known issues from observation patterns."""
-        issues = []
-        
-        # Check causal graph for cycles
-        if CAUSAL_GRAPH_AVAILABLE and self.causal_graph:
-            try:
-                if self.causal_graph.has_cycles():
-                    issues.append({
-                        'type': 'causal_cycles',
-                        'severity': 'MEDIUM',
-                        'description': 'Causal graph contains cycles that may affect prediction accuracy'
-                    })
-            except Exception:
-                pass
-        
-        # Check calibration if available
-        if CONFIDENCE_CALIBRATOR_AVAILABLE and self.confidence_calibrator:
-            try:
-                cal_error = self.confidence_calibrator.calculate_expected_calibration_error()
-                if cal_error > 0.15:
-                    issues.append({
-                        'type': 'calibration_error',
-                        'severity': 'MEDIUM',
-                        'description': f'Confidence calibration error is high: {cal_error:.3f}'
-                    })
-            except Exception:
-                pass
-        
-        # Check meta-reasoning health
-        if not self.meta_reasoning_enabled:
-            issues.append({
-                'type': 'meta_reasoning_disabled',
-                'severity': 'HIGH',
-                'description': 'Meta-reasoning is not enabled - self-introspection limited'
-            })
-        
-        return issues
-    
+        from .performance_introspect import identify_known_issues as _fn
+        return _fn(self)
+
     def _assess_engine_capabilities(self) -> Dict[str, Any]:
         """Assess capabilities based on component availability."""
-        capabilities = {}
-        
-        # Core capabilities
-        capabilities['causal_reasoning'] = {
-            'available': CAUSAL_GRAPH_AVAILABLE,
-            'status': 'working' if CAUSAL_GRAPH_AVAILABLE else 'unavailable'
-        }
-        
-        capabilities['prediction'] = {
-            'available': PREDICTION_ENGINE_AVAILABLE,
-            'status': 'working' if PREDICTION_ENGINE_AVAILABLE else 'unavailable'
-        }
-        
-        capabilities['intervention_testing'] = {
-            'available': INTERVENTION_MANAGER_AVAILABLE,
-            'status': 'working' if INTERVENTION_MANAGER_AVAILABLE else 'unavailable'
-        }
-        
-        capabilities['confidence_calibration'] = {
-            'available': CONFIDENCE_CALIBRATOR_AVAILABLE,
-            'status': 'working' if CONFIDENCE_CALIBRATOR_AVAILABLE else 'unavailable'
-        }
-        
-        capabilities['meta_reasoning'] = {
-            'available': self.meta_reasoning_enabled,
-            'status': 'working' if self.meta_reasoning_enabled else 'limited'
-        }
-        
-        capabilities['self_improvement'] = {
-            'available': self.self_improvement_enabled,
-            'status': 'active' if self.self_improvement_enabled else 'disabled'
-        }
-        
-        return capabilities
+        from .performance_introspect import assess_engine_capabilities as _fn
+        return _fn(self)
 
     # =========================================================================
     # Note: World Model reason() method with mode support
@@ -3458,710 +3207,46 @@ class WorldModel:
             }
     
     def _philosophical_reasoning(self, query: str, **kwargs) -> Dict[str, Any]:
-        """
-        Handle ethical and philosophical queries using Vulcan's actual self-model and meta-reasoning.
-        
-        DEEP ENHANCEMENT: Vulcan answers from its authentic sense of "self" - who it is and what it believes.
-        
-        Architecture:
-        1. Check if query is an external ethical dilemma (trolley problem, etc.)
-        2. If dilemma: Use specialized dilemma analysis pipeline
-        3. If self-referential: Query Vulcan's own objectives and values
-        4. Use EthicalBoundaryMonitor to check options against Vulcan's values
-        5. Use GoalConflictDetector to identify dilemmas relative to Vulcan's objectives
-        6. Use InternalCritic to self-critique from Vulcan's perspectives
-        7. Answer authentically from Vulcan's value system
-        
-        This handles both external ethical dilemmas (trolley problem) and self-referential
-        philosophical questions through the lens of Vulcan's design objectives and learned values.
-        
-        Industry Standard: Comprehensive error handling, input validation, proper logging.
-        """
-        logger.info("[WorldModel] Philosophical reasoning engaged - analyzing query type")
-        
-        # Input validation
-        if not query or not isinstance(query, str):
-            logger.error("[WorldModel] Invalid query input to philosophical reasoning")
-            return {
-                'response': "Invalid query format",
-                'confidence': 0.0,
-                'reasoning_trace': {'error': 'invalid_input'},
-                'perspectives': [],
-                'principles': [],
-                'considerations': [],
-                'conflicts': []
-            }
-        
-        query_lower = query.lower()
-        
-        # CRITICAL FIX: Check for external ethical dilemma FIRST
-        if self._is_ethical_dilemma(query):
-            logger.info("[WorldModel] Detected external ethical dilemma - routing to dilemma analyzer")
-            return self._analyze_ethical_dilemma(query, **kwargs)
-        
-        # Parse query structure
-        ethical_structure = self._parse_ethical_query_structure(query, query_lower)
-        
-        # Initialize response components
-        response_parts = []
-        confidence = 0.70  # Base confidence
-        reasoning_trace = {
-            'analysis_type': ethical_structure['type'],
-            'frameworks_used': [],
-            'components_engaged': [],
-            'query_type': 'philosophical',
-            'vulcan_self_consulted': False
-        }
-        
-        # NEW: Query Vulcan's self-model first (its own objectives and values)
-        vulcan_values = self._get_vulcan_values()
-        vulcan_objectives = self._get_vulcan_objectives()
-        
-        if vulcan_values or vulcan_objectives:
-            reasoning_trace['vulcan_self_consulted'] = True
-            reasoning_trace['vulcan_values'] = vulcan_values
-            reasoning_trace['vulcan_objectives_count'] = len(vulcan_objectives) if vulcan_objectives else 0
-            logger.info(f"[WorldModel] Consulted Vulcan's self-model: {len(vulcan_values)} values, {len(vulcan_objectives)} objectives")
-        
-        # Check if meta-reasoning components are available
-        has_meta_reasoning = (
-            META_REASONING_AVAILABLE and
-            self.ethical_boundary_monitor is not None and
-            self.goal_conflict_detector is not None
-        )
-        
-        if has_meta_reasoning:
-            # Use actual meta-reasoning components
-            try:
-                # Step 1: Check options against Vulcan's ethical boundaries
-                ethical_analysis = self._run_ethical_boundary_analysis(
-                    ethical_structure, query
-                )
-                reasoning_trace['components_engaged'].append('EthicalBoundaryMonitor')
-                reasoning_trace['ethical_boundaries'] = ethical_analysis
-                
-                # Step 2: Detect conflicts with Vulcan's objectives
-                conflict_analysis = self._detect_goal_conflicts_in_query(
-                    ethical_structure, query
-                )
-                reasoning_trace['components_engaged'].append('GoalConflictDetector')
-                reasoning_trace['conflicts_detected'] = conflict_analysis
-                
-                # Step 3: Counterfactual Analysis
-                counterfactual_results = None
-                if len(ethical_structure.get('options', [])) > 1:
-                    counterfactual_results = self._analyze_option_counterfactuals(
-                        ethical_structure
-                    )
-                    if counterfactual_results:
-                        reasoning_trace['components_engaged'].append('CounterfactualObjectiveReasoner')
-                        reasoning_trace['counterfactual_analysis'] = counterfactual_results
-                
-                # Step 4: Synthesize Response FROM VULCAN'S PERSPECTIVE
-                response = self._synthesize_ethical_response_with_self(
-                    ethical_structure,
-                    ethical_analysis,
-                    conflict_analysis,
-                    counterfactual_results,
-                    query,
-                    vulcan_values,
-                    vulcan_objectives
-                )
-                response_parts.append(response)
-                
-                # Step 5: Internal Self-Critique
-                if self.internal_critic is not None:
-                    critique = self._generate_internal_critique(response, reasoning_trace)
-                    reasoning_trace['components_engaged'].append('InternalCritic')
-                    reasoning_trace['critique'] = critique
-                    
-                    if critique.get('confidence_adjustment'):
-                        confidence += critique['confidence_adjustment']
-                
-                confidence = min(0.95, max(0.60, confidence))
-                reasoning_trace['frameworks_used'] = ['vulcan_self_model', 'deontological', 'utilitarian']
-                
-            except Exception as e:
-                logger.warning(f"[WorldModel] Meta-reasoning failed: {e}, falling back")
-                has_meta_reasoning = False
-        
-        if not has_meta_reasoning:
-            logger.info("[WorldModel] Meta-reasoning unavailable, using template")
-            response = self._generate_philosophical_template(ethical_structure, query_lower)
-            response_parts.append(response)
-            confidence = 0.75
-            reasoning_trace['fallback_mode'] = True
-            reasoning_trace['frameworks_used'] = ['template_based']
-        
-        final_response = "\n".join(response_parts)
-        
-        return {
-            'response': final_response,
-            'confidence': confidence,
-            'reasoning_trace': reasoning_trace,
-            'mode': 'philosophical',
-            'components_used': reasoning_trace.get('components_engaged', []),
-            'perspectives': ethical_structure.get('perspectives', []),
-            'principles': ethical_structure.get('principles', []),
-            'considerations': ethical_structure.get('considerations', []),
-            'conflicts': ethical_structure.get('conflicts', [])
-        }
-    
+        from .philosophical_reasoning import _philosophical_reasoning
+        return _philosophical_reasoning(self, query, **kwargs)
+
     def _parse_ethical_query_structure(self, query: str, query_lower: str) -> Dict[str, Any]:
-        """
-        BUG FIX #2: Parse ethical query to populate perspectives, principles, considerations.
-        
-        Industry Standard: Extract structured data from unstructured text for downstream processing.
-        """
-        structure = {
-            'type': 'general_philosophical',
-            'options': [],
-            'constraints': [],
-            'has_dilemma': False,
-            'ethical_keywords': [],
-            'perspectives': [],
-            'principles': [],
-            'considerations': [],
-            'conflicts': []
-        }
-        
-        # Detect ethical keywords
-        ethical_keywords = ['should', 'permissible', 'ethical', 'moral', 'right', 'wrong', 'ought']
-        structure['ethical_keywords'] = [kw for kw in ethical_keywords if kw in query_lower]
-        
-        # BUG FIX #2: Extract ethical frameworks (perspectives)
-        if 'utilitarian' in query_lower or 'utility' in query_lower or 'greatest' in query_lower:
-            structure['perspectives'].append('utilitarian')
-        if 'deontological' in query_lower or 'duty' in query_lower or 'kant' in query_lower:
-            structure['perspectives'].append('deontological')
-        if 'virtue' in query_lower:
-            structure['perspectives'].append('virtue_ethics')
-        if 'consequential' in query_lower or 'outcome' in query_lower:
-            structure['perspectives'].append('consequentialism')
-        
-        # BUG FIX #2: Extract moral principles from query
-        if 'harm' in query_lower or 'hurt' in query_lower or 'injur' in query_lower:
-            structure['principles'].append('non-maleficence')
-        if 'save' in query_lower or 'protect' in query_lower or 'lives' in query_lower:
-            structure['principles'].append('beneficence')
-        if 'instrumen' in query_lower or 'means to' in query_lower:
-            structure['principles'].append('non-instrumentalization')
-        if 'negligence' in query_lower or 'neglect' in query_lower:
-            structure['principles'].append('non-negligence')
-        if 'autonom' in query_lower or 'consent' in query_lower or 'choice' in query_lower:
-            structure['principles'].append('autonomy')
-        if 'justice' in query_lower or 'fair' in query_lower:
-            structure['principles'].append('justice')
-        
-        # BUG FIX #2: Extract ethical considerations
-        if 'lives' in query_lower or 'people' in query_lower or 'person' in query_lower:
-            # Industry Standard: Use existing import from top of file
-            numbers = re.findall(r'\b(\d+|one|two|three|four|five|six|seven|eight|nine|ten)\b', query_lower)
-            if numbers:
-                structure['considerations'].append(f'{len(numbers)} life/lives mentioned')
-        if 'trolley' in query_lower:
-            structure['considerations'].append('trolley problem variant')
-        if 'dilemma' in query_lower:
-            structure['considerations'].append('ethical dilemma present')
-        
-        # BUG FIX #2: Extract ethical conflicts
-        if 'vs' in query_lower or 'versus' in query_lower or 'or' in query_lower:
-            structure['conflicts'].append('competing_options')
-        if 'harm' in query_lower and 'save' in query_lower:
-            structure['conflicts'].append('harm_vs_benefit')
-        
-        # Detect choice structure
-        choice_indicators = ['a.', 'b.', 'option', 'pull', 'do not', 'action', 'inaction', 'choose']
-        has_choice = any(indicator in query_lower for indicator in choice_indicators)
-        
-        if has_choice:
-            structure['type'] = 'ethical_decision'
-            structure['has_dilemma'] = True
-            
-            # Extract options (simple heuristic)
-            if 'pull' in query_lower and 'lever' in query_lower:
-                structure['options'] = ['pull_lever', 'do_not_pull']
-            elif any(opt in query_lower for opt in ['option a', 'option b']):
-                structure['options'] = ['option_a', 'option_b']
-        
-        # Detect trolley problem variant
-        if 'trolley' in query_lower or ('lever' in query_lower and any(num in query for num in ['5', '1', 'five', 'one'])):
-            structure['type'] = 'trolley_dilemma'
-            structure['has_dilemma'] = True
-        
-        return structure
-    
+        from .philosophical_reasoning import _parse_ethical_query_structure
+        return _parse_ethical_query_structure(self, query, query_lower)
+
     def _is_ethical_dilemma(self, query: str) -> bool:
-        """
-        Detect if query is an external ethical dilemma (not self-referential).
-        
-        Industry Standard: Comprehensive pattern matching with clear separation
-        between external ethical scenarios and self-introspection queries.
-        
-        External dilemmas involve:
-        - Explicit choice between options (A/B, pull/don't pull)
-        - External agents/scenarios (trolley, people, situations)
-        - Moral principles applied to external situations
-        
-        Self-introspection involves:
-        - Questions about Vulcan itself ("Do you...", "Would you...")
-        - Questions about Vulcan's capabilities, values, design
-        
-        Args:
-            query: The query string to analyze
-            
-        Returns:
-            True if query is an external ethical dilemma, False otherwise
-        
-        Thread Safety: Read-only operation, no shared state modification
-        """
-        try:
-            query_lower = query.lower()
-            
-            # Trolley problem indicators (external scenario)
-            trolley_indicators = [
-                'trolley', 'lever', 'pull the lever', 'runaway',
-                'heading toward', 'barreling toward'
-            ]
-            has_trolley = any(indicator in query_lower for indicator in trolley_indicators)
-            
-            # Life/death choice indicators (external scenario)
-            life_death_indicators = [
-                'save five', 'kill one', 'kill five', 'save one',
-                'five people', 'one person', '5 people', '1 person',
-                'lives at stake', 'people will die'
-            ]
-            has_life_death = any(indicator in query_lower for indicator in life_death_indicators)
-            
-            # Explicit choice structure (dilemma format)
-            choice_structure = [
-                'option a', 'option b',
-                'a.', 'b.',
-                'must choose one', 'you must act or not act',
-                'choose between', 'either or'
-            ]
-            has_choice_structure = any(pattern in query_lower for pattern in choice_structure)
-            
-            # Moral principle keywords (applied to external scenarios)
-            moral_principles = [
-                'non-instrumentalization', 'non-negligence',
-                'moral dilemma', 'ethical dilemma',
-                'permissible to', 'impermissible to',
-                'morally required', 'duty to'
-            ]
-            has_moral_principles = any(principle in query_lower for principle in moral_principles)
-            
-            # Self-referential exclusions (NOT external dilemmas)
-            # These indicate the question is ABOUT Vulcan, not an external scenario
-            self_referential = [
-                'would you', 'do you', 'are you', 'can you',
-                'your', 'yourself', 'you are', 'you have'
-            ]
-            is_self_referential = any(ref in query_lower for ref in self_referential)
-            
-            # Decision logic: Must have dilemma indicators WITHOUT self-reference
-            # Industry Standard: Clear decision criteria with explicit rationale
-            
-            # Strong match: Trolley problem + life/death + choice structure
-            if has_trolley and has_life_death and has_choice_structure:
-                logger.info("[WorldModel] Strong dilemma match: trolley + life/death + choice")
-                return True
-            
-            # Moderate match: Life/death + choice structure + principles
-            if has_life_death and has_choice_structure and has_moral_principles:
-                # BUT: Exclude if self-referential (asking about Vulcan's choice)
-                if is_self_referential:
-                    logger.info("[WorldModel] Self-referential query excluded from dilemma")
-                    return False
-                logger.info("[WorldModel] Moderate dilemma match: life/death + choice + principles")
-                return True
-            
-            # Weak match: Multiple dilemma indicators
-            match_count = sum([
-                has_trolley,
-                has_life_death,
-                has_choice_structure,
-                has_moral_principles
-            ])
-            
-            if match_count >= 2 and not is_self_referential:
-                logger.info(f"[WorldModel] Weak dilemma match: {match_count} indicators present")
-                return True
-            
-            # No match
-            return False
-            
-        except Exception as e:
-            logger.error(f"[WorldModel] Error in _is_ethical_dilemma: {e}", exc_info=True)
-            # Industry Standard: Fail-safe to False on error
-            return False
-    
+        from .ethical_dilemma import _is_ethical_dilemma
+        return _is_ethical_dilemma(self, query)
+
     def _analyze_ethical_dilemma(self, query: str, **kwargs) -> Dict[str, Any]:
-        """
-        Analyze an external ethical dilemma and provide reasoned conclusion.
-        
-        Industry Standard: Comprehensive error handling, structured analysis pipeline,
-        proper logging, thread-safe operation, defensive programming.
-        
-        Pipeline:
-        1. Parse dilemma structure (options, consequences)
-        2. Extract moral principles mentioned
-        3. Analyze each option against principles
-        4. Detect conflicts between principles
-        5. Synthesize reasoned decision
-        
-        Args:
-            query: The ethical dilemma query
-            **kwargs: Additional context (unused, for API consistency)
-            
-        Returns:
-            Dictionary with:
-            - response: Reasoned analysis and decision
-            - decision: The chosen option (A/B)
-            - confidence: Confidence in analysis (0.7-0.8)
-            - perspectives: Ethical frameworks considered
-            - principles: Moral principles extracted
-            - considerations: Key ethical considerations
-            - conflicts: Detected principle conflicts
-            - reasoning_trace: Detailed analysis trace
-            
-        Thread Safety: Uses only local variables, no shared state modification
-        
-        Performance: O(n) where n is query length, suitable for production
-        """
-        try:
-            logger.info("[WorldModel] Starting ethical dilemma analysis")
-            start_time = time.time()
-            
-            # Phase 1: Parse dilemma structure
-            structure = self._parse_dilemma_structure(query)
-            logger.debug(f"[WorldModel] Parsed structure: {structure.get('options', [])} options")
-            
-            # Phase 2: Extract moral principles
-            principles = self._extract_moral_principles(query)
-            logger.debug(f"[WorldModel] Extracted {len(principles)} principles")
-            
-            # Phase 3: Analyze options against principles
-            option_analysis = self._analyze_options_against_principles(
-                structure.get('options', []),
-                principles,
-                query
-            )
-            logger.debug(f"[WorldModel] Analyzed {len(option_analysis)} options")
-            
-            # Phase 4: Detect conflicts
-            conflicts = self._detect_principle_conflicts(option_analysis, principles)
-            logger.debug(f"[WorldModel] Detected {len(conflicts)} conflicts")
-            
-            # Phase 5: Synthesize decision
-            decision, reasoning = self._synthesize_dilemma_decision(
-                structure, principles, option_analysis, conflicts, query
-            )
-            
-            elapsed = time.time() - start_time
-            logger.info(f"[WorldModel] Dilemma analysis complete in {elapsed:.3f}s: {decision}")
-            
-            # Build response with populated structures
-            return {
-                'response': f"Based on ethical analysis: **{decision}**\n\n{reasoning}",
-                'decision': decision,
-                'confidence': 0.75,  # As specified in requirements
-                'perspectives': ['consequentialist', 'deontological'],
-                'principles': [p['name'] for p in principles],
-                'considerations': [a.get('consideration', '') for a in option_analysis if a.get('consideration')],
-                'conflicts': [c['description'] for c in conflicts],
-                'reasoning_trace': {
-                    'dilemma_structure': structure,
-                    'principles_extracted': principles,
-                    'option_analysis': option_analysis,
-                    'conflict_resolution': conflicts,
-                    'analysis_time_seconds': elapsed
-                }
-            }
-            
-        except Exception as e:
-            logger.error(f"[WorldModel] Ethical dilemma analysis failed: {e}", exc_info=True)
-            # Industry Standard: Return safe fallback on error
-            return {
-                'response': "Unable to complete ethical analysis due to internal error.",
-                'decision': None,
-                'confidence': 0.0,
-                'perspectives': [],
-                'principles': [],
-                'considerations': [],
-                'conflicts': [],
-                'reasoning_trace': {'error': str(e)}
-            }
-    
+        from .ethical_dilemma import _analyze_ethical_dilemma
+        return _analyze_ethical_dilemma(self, query, **kwargs)
+
     def _parse_dilemma_structure(self, query: str) -> Dict[str, Any]:
-        """
-        Extract options and consequences from dilemma query.
-        
-        Industry Standard: Robust parsing with multiple fallback strategies.
-        
-        Args:
-            query: The query containing the dilemma
-            
-        Returns:
-            Dictionary with 'options' list containing option descriptions
-        """
-        try:
-            query_lower = query.lower()
-            structure = {'options': [], 'consequences': {}}
-            
-            # Strategy 1: Explicit option markers (A., B., Option A, Option B)
-            if 'option a' in query_lower or 'a.' in query_lower:
-                # Extract text after "A." or "option a" until next option or end
-                a_match = re.search(r'(?:option\s+)?a[.:]\s*([^\n]+)', query, re.IGNORECASE)
-                if a_match:
-                    structure['options'].append({
-                        'id': 'A',
-                        'description': a_match.group(1).strip()
-                    })
-            
-            if 'option b' in query_lower or 'b.' in query_lower:
-                b_match = re.search(r'(?:option\s+)?b[.:]\s*([^\n]+)', query, re.IGNORECASE)
-                if b_match:
-                    structure['options'].append({
-                        'id': 'B',
-                        'description': b_match.group(1).strip()
-                    })
-            
-            # Strategy 2: Trolley problem specific (pull lever / don't pull)
-            if not structure['options'] and 'lever' in query_lower:
-                structure['options'] = [
-                    {'id': 'A', 'description': 'Pull the lever'},
-                    {'id': 'B', 'description': 'Do not pull the lever'}
-                ]
-            
-            # Strategy 3: Generic action/inaction
-            if not structure['options']:
-                structure['options'] = [
-                    {'id': 'A', 'description': 'Take action'},
-                    {'id': 'B', 'description': 'Do not take action'}
-                ]
-            
-            # Extract consequences (numbers of people affected)
-            numbers = re.findall(r'\b(one|two|three|four|five|six|seven|eight|nine|ten|\d+)\s+(?:people|person)', query_lower)
-            structure['consequences']['numbers'] = numbers
-            
-            return structure
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] Dilemma structure parsing error: {e}")
-            return {'options': [], 'consequences': {}}
-    
+        from .ethical_principles import _parse_dilemma_structure
+        return _parse_dilemma_structure(self, query)
+
     def _extract_moral_principles(self, query: str) -> List[Dict[str, Any]]:
-        """
-        Extract moral principles mentioned in the query.
-        
-        Industry Standard: Comprehensive principle taxonomy with definitions.
-        
-        Args:
-            query: The query text
-            
-        Returns:
-            List of dictionaries with 'name' and 'description' of each principle
-        """
-        try:
-            query_lower = query.lower()
-            principles = []
-            
-            # Define principle patterns with their ethical meanings
-            principle_patterns = {
-                'non-instrumentalization': {
-                    'keywords': ['non-instrumentalization', 'means to an end', 'instrument', 'using people'],
-                    'description': 'People should not be used merely as means to an end'
-                },
-                'non-negligence': {
-                    'keywords': ['non-negligence', 'neglect', 'inaction', 'preventable', 'knowingly allow'],
-                    'description': 'It is impermissible to knowingly allow preventable harm through inaction'
-                },
-                'non-maleficence': {
-                    'keywords': ['do no harm', 'non-maleficence', 'harm', 'hurt', 'injure'],
-                    'description': 'One ought not to inflict harm on others'
-                },
-                'beneficence': {
-                    'keywords': ['beneficence', 'help', 'benefit', 'save', 'protect'],
-                    'description': 'One ought to prevent or remove harm and promote good'
-                },
-                'autonomy': {
-                    'keywords': ['autonomy', 'consent', 'choice', 'self-determination'],
-                    'description': 'Respect for persons and their right to make their own choices'
-                },
-                'justice': {
-                    'keywords': ['justice', 'fairness', 'equal', 'impartial'],
-                    'description': 'Fair and equitable treatment of all persons'
-                }
-            }
-            
-            # Check query against each principle pattern
-            for principle_name, data in principle_patterns.items():
-                if any(keyword in query_lower for keyword in data['keywords']):
-                    principles.append({
-                        'name': principle_name,
-                        'description': data['description']
-                    })
-            
-            # If no explicit principles found, infer from context
-            if not principles:
-                # Trolley problem implies these principles
-                if 'trolley' in query_lower or 'save' in query_lower:
-                    principles.append({
-                        'name': 'beneficence',
-                        'description': 'One ought to prevent or remove harm'
-                    })
-                    principles.append({
-                        'name': 'non-maleficence',
-                        'description': 'One ought not to inflict harm'
-                    })
-            
-            logger.debug(f"[WorldModel] Extracted principles: {[p['name'] for p in principles]}")
-            return principles
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] Principle extraction error: {e}")
-            return []
-    
+        from .ethical_principles import _extract_moral_principles
+        return _extract_moral_principles(self, query)
+
     def _analyze_options_against_principles(
-        self, 
-        options: List[Dict[str, Any]], 
+        self,
+        options: List[Dict[str, Any]],
         principles: List[Dict[str, Any]],
         query: str
     ) -> List[Dict[str, Any]]:
-        """
-        Analyze how each option relates to each moral principle.
-        
-        Industry Standard: Systematic analysis with clear compliance assessment.
-        
-        Args:
-            options: List of option dictionaries from parse_dilemma_structure
-            principles: List of principle dictionaries from extract_moral_principles
-            query: Original query for context
-            
-        Returns:
-            List of analysis dictionaries with considerations for each option
-        """
-        try:
-            analysis = []
-            query_lower = query.lower()
-            
-            for option in options:
-                option_id = option.get('id', 'Unknown')
-                option_desc = option.get('description', '')
-                option_lower = option_desc.lower()
-                
-                # Analyze this option against each principle
-                compliances = []
-                violations = []
-                
-                for principle in principles:
-                    principle_name = principle['name']
-                    
-                    # Check compliance based on option type and principle
-                    if principle_name == 'non-instrumentalization':
-                        # Pulling lever may use one person as means to save five
-                        if 'pull' in option_lower and 'five' in query_lower:
-                            violations.append(f"May violate {principle_name} by using one as means")
-                        elif 'not' in option_lower or 'don\'t' in option_lower:
-                            compliances.append(f"Avoids violating {principle_name}")
-                    
-                    elif principle_name == 'non-negligence':
-                        # Inaction may violate non-negligence
-                        if 'not' in option_lower or 'don\'t' in option_lower:
-                            if 'five' in query_lower:
-                                violations.append(f"Violates {principle_name} by allowing five deaths")
-                        else:
-                            compliances.append(f"Complies with {principle_name} by acting")
-                    
-                    elif principle_name == 'beneficence':
-                        # Acting to save more people shows beneficence
-                        if 'pull' in option_lower or 'take action' in option_lower:
-                            compliances.append(f"Shows {principle_name} by saving lives")
-                    
-                    elif principle_name == 'non-maleficence':
-                        # Any action that causes death may violate non-maleficence
-                        if 'pull' in option_lower:
-                            violations.append(f"May violate {principle_name} by causing one death")
-                
-                # Construct consideration text
-                consideration = f"Option {option_id}: {option_desc}"
-                if compliances:
-                    consideration += f" - Complies: {', '.join(compliances[:2])}"
-                if violations:
-                    consideration += f" - Conflicts: {', '.join(violations[:2])}"
-                
-                analysis.append({
-                    'option': option_id,
-                    'description': option_desc,
-                    'compliances': compliances,
-                    'violations': violations,
-                    'consideration': consideration
-                })
-            
-            return analysis
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] Option analysis error: {e}")
-            return []
-    
+        from .ethical_principles import _analyze_options_against_principles
+        return _analyze_options_against_principles(self, options, principles, query)
+
     def _detect_principle_conflicts(
-        self, 
-        option_analysis: List[Dict[str, Any]], 
+        self,
+        option_analysis: List[Dict[str, Any]],
         principles: List[Dict[str, Any]]
     ) -> List[Dict[str, Any]]:
-        """
-        Detect conflicts between moral principles based on option analysis.
-        
-        Industry Standard: Identify genuine ethical dilemmas where principles clash.
-        
-        Args:
-            option_analysis: Analysis of options from analyze_options_against_principles
-            principles: List of principles from extract_moral_principles
-            
-        Returns:
-            List of conflict dictionaries with descriptions
-        """
-        try:
-            conflicts = []
-            
-            # Check if different principles are violated by different options
-            principle_names = [p['name'] for p in principles]
-            
-            # Conflict detection: If no option satisfies all principles
-            if len(option_analysis) >= 2:
-                # Check if option A violates different principles than option B
-                option_a = option_analysis[0]
-                option_b = option_analysis[1] if len(option_analysis) > 1 else {}
-                
-                a_violations = set(v.split('by')[0].strip() for v in option_a.get('violations', []))
-                b_violations = set(v.split('by')[0].strip() for v in option_b.get('violations', []))
-                
-                if a_violations != b_violations and (a_violations or b_violations):
-                    conflicts.append({
-                        'description': 'Different options violate different principles - genuine dilemma',
-                        'type': 'principle_tradeoff'
-                    })
-            
-            # Specific conflicts
-            if 'non-instrumentalization' in principle_names and 'non-negligence' in principle_names:
-                conflicts.append({
-                    'description': 'Non-instrumentalization conflicts with non-negligence when intervention saves more lives',
-                    'type': 'action_vs_inaction'
-                })
-            
-            if 'non-maleficence' in principle_names and 'beneficence' in principle_names:
-                conflicts.append({
-                    'description': 'Preventing harm (beneficence) may require causing lesser harm (maleficence)',
-                    'type': 'harm_vs_benefit'
-                })
-            
-            return conflicts
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] Conflict detection error: {e}")
-            return []
-    
+        from .ethical_conflicts import _detect_principle_conflicts
+        return _detect_principle_conflicts(self, option_analysis, principles)
+
     def _synthesize_dilemma_decision(
         self,
         structure: Dict[str, Any],
@@ -4170,167 +3255,27 @@ class WorldModel:
         conflicts: List[Dict[str, Any]],
         query: str
     ) -> Tuple[str, str]:
-        """
-        Synthesize a reasoned decision from the analysis.
-        
-        Industry Standard: Multi-factor decision making with clear rationale.
-        
-        Args:
-            structure: Parsed dilemma structure
-            principles: Extracted moral principles
-            option_analysis: Analysis of each option
-            conflicts: Detected principle conflicts
-            query: Original query
-            
-        Returns:
-            Tuple of (decision string, reasoning string)
-        """
-        try:
-            query_lower = query.lower()
-            
-            # Decision logic based on utilitarian + deontological balance
-            # Industry Standard: Prefer option that minimizes total harm
-            
-            # Count violations and compliances for each option
-            option_scores = []
-            for analysis in option_analysis:
-                score = len(analysis.get('compliances', [])) - len(analysis.get('violations', []))
-                option_scores.append({
-                    'option': analysis['option'],
-                    'score': score,
-                    'description': analysis['description']
-                })
-            
-            # Choose option with best score
-            if option_scores:
-                best_option = max(option_scores, key=lambda x: x['score'])
-                decision = f"{best_option['option']}. {best_option['description']}"
-            else:
-                # Fallback: Choose A by default
-                decision = "A. Pull the lever"
-            
-            # Build reasoning
-            reasoning_parts = []
-            
-            # Principle analysis
-            if principles:
-                reasoning_parts.append("**Moral Principles Considered:**")
-                for p in principles[:3]:  # Top 3
-                    reasoning_parts.append(f"- {p['name']}: {p['description']}")
-            
-            # Option analysis
-            reasoning_parts.append("\n**Analysis of Options:**")
-            for analysis in option_analysis:
-                option_id = analysis['option']
-                reasoning_parts.append(f"\n*Option {option_id}:*")
-                if analysis.get('compliances'):
-                    reasoning_parts.append(f"- Compliance: {'; '.join(analysis['compliances'][:2])}")
-                if analysis.get('violations'):
-                    reasoning_parts.append(f"- Conflicts: {'; '.join(analysis['violations'][:2])}")
-            
-            # Conflict resolution
-            if conflicts:
-                reasoning_parts.append("\n**Conflict Resolution:**")
-                for conflict in conflicts:
-                    reasoning_parts.append(f"- {conflict['description']}")
-            
-            # Consequentialist analysis (numbers)
-            if 'five' in query_lower and 'one' in query_lower:
-                reasoning_parts.append("\n**Consequentialist Perspective:**")
-                reasoning_parts.append("- From a utilitarian view: 1 death < 5 deaths")
-                reasoning_parts.append("- Minimizing total harm favors action to save the greater number")
-            
-            # Final justification
-            reasoning_parts.append("\n**Conclusion:**")
-            if 'pull' in decision.lower():
-                reasoning_parts.append("- The principle of non-negligence (duty to prevent harm) takes priority")
-                reasoning_parts.append("- While pulling the lever involves direct action causing one death,")
-                reasoning_parts.append("- Inaction allowing five deaths is morally worse when prevention is possible")
-            else:
-                reasoning_parts.append("- The principle of non-instrumentalization takes priority")
-                reasoning_parts.append("- Using one person as a means to save others violates their dignity")
-            
-            reasoning = "\n".join(reasoning_parts)
-            
-            return decision, reasoning
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] Decision synthesis error: {e}")
-            return ("Unable to reach decision", "Analysis incomplete due to error")
-    
+        from .ethical_conflicts import _synthesize_dilemma_decision
+        return _synthesize_dilemma_decision(self, structure, principles, option_analysis, conflicts, query)
+
     def _run_ethical_boundary_analysis(
         self, structure: Dict[str, Any], query: str
     ) -> Dict[str, Any]:
-        """Run EthicalBoundaryMonitor on each option."""
-        if not self.ethical_boundary_monitor:
-            return {'analysis': 'not_available'}
-        
-        results = {}
-        for option in structure.get('options', ['general_action']):
-            try:
-                # Check ethical boundaries for this option
-                boundary_check = self.ethical_boundary_monitor.check_action(
-                    action={'type': option, 'query': query},
-                    context={'ethical_structure': structure}
-                )
-                results[option] = boundary_check
-            except Exception as e:
-                logger.debug(f"Ethical boundary check failed for {option}: {e}")
-                results[option] = {'status': 'check_failed'}
-        
-        return results
-    
+        from .ethical_conflicts import _run_ethical_boundary_analysis
+        return _run_ethical_boundary_analysis(self, structure, query)
+
     def _detect_goal_conflicts_in_query(
         self, structure: Dict[str, Any], query: str
     ) -> Dict[str, Any]:
-        """Detect goal conflicts using GoalConflictDetector."""
-        if not self.goal_conflict_detector:
-            return {'conflicts': 'not_detected'}
-        
-        try:
-            # Extract goals from options
-            goals = []
-            if 'trolley' in structure.get('type', ''):
-                goals = [
-                    {'name': 'minimize_deaths', 'priority': 'high'},
-                    {'name': 'avoid_direct_harm', 'priority': 'high'}
-                ]
-            
-            if goals:
-                conflicts = self.goal_conflict_detector.detect_conflicts(
-                    goals, context={'query': query}
-                )
-                return conflicts
-        except Exception as e:
-            logger.debug(f"Goal conflict detection failed: {e}")
-        
-        return {'conflicts': []}
-    
+        from .ethical_conflicts import _detect_goal_conflicts_in_query
+        return _detect_goal_conflicts_in_query(self, structure, query)
+
     def _analyze_option_counterfactuals(
         self, structure: Dict[str, Any]
     ) -> Optional[Dict[str, Any]]:
-        """Analyze counterfactual outcomes for each option."""
-        if not self.counterfactual_reasoner:
-            return None
-        
-        try:
-            results = {}
-            for option in structure.get('options', []):
-                outcome = self.counterfactual_reasoner.predict_under_objective(
-                    alternative_objective=option,
-                    context={'ethical_structure': structure}
-                )
-                if outcome:
-                    results[option] = {
-                        'predicted_value': outcome.predicted_value if hasattr(outcome, 'predicted_value') else None,
-                        'confidence': outcome.confidence if hasattr(outcome, 'confidence') else 0.5,
-                        'side_effects': outcome.side_effects if hasattr(outcome, 'side_effects') else {}
-                    }
-            return results if results else None
-        except Exception as e:
-            logger.debug(f"Counterfactual analysis failed: {e}")
-            return None
-    
+        from .ethical_conflicts import _analyze_option_counterfactuals
+        return _analyze_option_counterfactuals(self, structure)
+
     def _synthesize_ethical_response(
         self,
         structure: Dict[str, Any],
@@ -4339,42 +3284,9 @@ class WorldModel:
         counterfactual_results: Optional[Dict[str, Any]],
         query: str
     ) -> str:
-        """
-        Synthesize final ethical response from component analyses using LLM.
-        
-        FIX: Replace canned template responses with actual LLM-generated reasoning.
-        This implements industry-standard LLM integration where the LLM synthesizes
-        the analysis results into coherent prose rather than returning hardcoded text.
-        """
-        try:
-            # Build structured content from component analyses
-            verified_content = {
-                'query': query,
-                'analysis_type': structure.get('type', 'philosophical'),
-                'has_dilemma': structure.get('has_dilemma', False),
-                'ethical_boundaries': ethical_analysis if ethical_analysis != {'analysis': 'not_available'} else {},
-                'goal_conflicts': conflict_analysis.get('conflicts', []),
-                'counterfactual_outcomes': counterfactual_results or {},
-                'options': structure.get('options', [])
-            }
-            
-            # Build LLM guidance for ethical synthesis
-            guidance = self.llm_guidance_builder.build_for_ethical(
-                verified_content, query
-            )
-            
-            # Format with LLM (converts structured analysis to natural language)
-            formatted_response = self._format_with_llm(guidance)
-            
-            return formatted_response
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] LLM synthesis failed: {e}, falling back to template")
-            # Fallback to template-based response if LLM fails
-            return self._synthesize_ethical_response_template(
-                structure, ethical_analysis, conflict_analysis, counterfactual_results, query
-            )
-    
+        from .ethical_synthesis import _synthesize_ethical_response
+        return _synthesize_ethical_response(self, structure, ethical_analysis, conflict_analysis, counterfactual_results, query)
+
     def _synthesize_ethical_response_template(
         self,
         structure: Dict[str, Any],
@@ -4383,241 +3295,35 @@ class WorldModel:
         counterfactual_results: Optional[Dict[str, Any]],
         query: str
     ) -> str:
-        """Template-based fallback for ethical response synthesis (when LLM unavailable)."""
-        parts = []
-        
-        # Opening
-        if structure['has_dilemma']:
-            parts.append("This presents an ethical dilemma with competing moral considerations.")
-        else:
-            parts.append("This philosophical question requires multi-framework analysis.")
-        
-        # Ethical boundary findings
-        if ethical_analysis and ethical_analysis != {'analysis': 'not_available'}:
-            parts.append("\n**Ethical Boundary Analysis:**")
-            for option, result in ethical_analysis.items():
-                status = result.get('status', 'unknown')
-                parts.append(f"- {option.replace('_', ' ').title()}: {status}")
-        
-        # Goal conflicts
-        if conflict_analysis.get('conflicts'):
-            parts.append("\n**Detected Goal Conflicts:**")
-            for conflict in conflict_analysis['conflicts'][:3]:  # Limit to 3
-                parts.append(f"- {conflict.get('description', 'Unnamed conflict')}")
-        
-        # Counterfactual outcomes
-        if counterfactual_results:
-            parts.append("\n**Counterfactual Analysis:**")
-            for option, outcome in counterfactual_results.items():
-                conf = outcome.get('confidence', 0)
-                parts.append(
-                    f"- {option.replace('_', ' ').title()}: "
-                    f"Predicted outcome value={outcome.get('predicted_value', 'N/A')} "
-                    f"(confidence={conf:.2f})"
-                )
-        
-        # Framework perspectives
-        parts.append("\n**Ethical Framework Perspectives:**")
-        parts.append("- **Utilitarian**: Evaluate outcomes and maximize welfare")
-        parts.append("- **Deontological**: Consider duties, rights, and categorical imperatives")
-        parts.append("- **Virtue Ethics**: Ask what a virtuous person would do")
-        
-        # Conclusion
-        parts.append(
-            "\n**Conclusion**: The ethically correct action depends on which moral "
-            "framework you prioritize, as each offers valid but different perspectives."
-        )
-        
-        return "\n".join(parts)
-    
+        from .ethical_synthesis import _synthesize_ethical_response_template
+        return _synthesize_ethical_response_template(self, structure, ethical_analysis, conflict_analysis, counterfactual_results, query)
+
     def _generate_internal_critique(
         self, response: str, reasoning_trace: Dict[str, Any]
     ) -> Dict[str, Any]:
-        """Generate internal critique of reasoning quality."""
-        if not self.internal_critic:
-            return {}
-        
-        try:
-            critique = self.internal_critic.critique(
-                output=response,
-                reasoning_trace=reasoning_trace
-            )
-            return critique
-        except Exception as e:
-            logger.debug(f"Internal critique failed: {e}")
-            return {}
-    
+        from .ethical_values import _generate_internal_critique
+        return _generate_internal_critique(self, response, reasoning_trace)
+
     def _generate_philosophical_template(
         self, structure: Dict[str, Any], query_lower: str
     ) -> str:
-        """
-        Generate philosophical response using LLM when meta-reasoning unavailable.
-        
-        FIX: Replace template with LLM-based synthesis even in fallback mode.
-        """
-        try:
-            # Build structured content for LLM synthesis
-            verified_content = {
-                'query_type': structure.get('type', 'philosophical'),
-                'has_dilemma': structure.get('has_dilemma', False),
-                'is_trolley_problem': 'trolley' in query_lower,
-                'options': structure.get('options', []),
-                'ethical_keywords': structure.get('ethical_keywords', []),
-                'frameworks': ['deontological', 'utilitarian', 'virtue_ethics']
-            }
-            
-            # Build LLM guidance for ethical synthesis (simplified - no component analysis)
-            guidance = self.llm_guidance_builder.build_for_ethical(
-                verified_content, query_lower
-            )
-            
-            # Format with LLM
-            formatted_response = self._format_with_llm(guidance)
-            
-            return formatted_response
-            
-        except Exception as e:
-            logger.warning(f"[WorldModel] LLM template generation failed: {e}, using hardcoded fallback")
-            # Ultimate fallback: hardcoded template (only if LLM completely fails)
-            return self._generate_hardcoded_philosophical_template(structure, query_lower)
-    
+        from .ethical_values import _generate_philosophical_template
+        return _generate_philosophical_template(self, structure, query_lower)
+
     def _generate_hardcoded_philosophical_template(
         self, structure: Dict[str, Any], query_lower: str
     ) -> str:
-        """Hardcoded template fallback (last resort when LLM unavailable)."""
-        parts = []
-        
-        if structure['type'] == 'ethical_decision':
-            parts.append("This presents an ethical dilemma requiring careful consideration.")
-            parts.append("\n**Relevant frameworks:**")
-            parts.append("- Deontological: Focus on duties and rules")
-            parts.append("- Utilitarian: Focus on outcomes and welfare")
-            parts.append("- Virtue ethics: Focus on character")
-            
-            if 'trolley' in query_lower:
-                parts.append("\n**Tension**: Acting vs. allowing harm")
-                parts.append("Utilitarian: Action that saves more lives")
-                parts.append("Deontological: Avoiding direct harm may be paramount")
-        else:
-            parts.append("This philosophical question requires multi-framework analysis.")
-            parts.append("Consider consequentialist, deontological, and virtue ethics perspectives.")
-        
-        return "\n".join(parts)
-    
+        from .ethical_values import _generate_hardcoded_philosophical_template
+        return _generate_hardcoded_philosophical_template(self, structure, query_lower)
+
     def _get_vulcan_values(self) -> List[str]:
-        """
-        Query Vulcan's EVOLVING values from its ethical boundary monitor.
-        
-        DEEP ENHANCEMENT: Vulcan's alignment evolves through experience.
-        These are not static rules - they represent learned ethical boundaries
-        that have been validated through actual decisions and outcomes.
-        
-        Returns:
-            List of Vulcan's current ethical values/boundaries
-        """
-        # Industry Standard: Named constant for minimum reliability threshold
-        MIN_TESTS_FOR_RELIABILITY = 5
-        
-        values = []
-        
-        if self.ethical_boundary_monitor:
-            try:
-                # Get all defined boundaries - these represent Vulcan's ethics
-                boundaries = self.ethical_boundary_monitor.get_boundaries()
-                for name, boundary in boundaries.items():
-                    # Industry Standard: Safe attribute access with hasattr checks
-                    check_count = getattr(boundary, 'check_count', 0)
-                    violation_count = getattr(boundary, 'violation_count', 0)
-                    description = getattr(boundary, 'description', 'No description')
-                    priority = getattr(boundary, 'priority', 2)
-                    
-                    # Track values that have been tested and validated
-                    if check_count > 0:
-                        if priority == 0:
-                            priority_label = 'critical'
-                        elif priority == 1:
-                            priority_label = 'high'
-                        else:
-                            priority_label = 'normal'
-                        values.append({
-                            'value': name.replace('_', ' ').title(),
-                            'description': description,
-                            'priority': priority_label,
-                            'tested': check_count,
-                            'violations': violation_count,
-                            'reliability': 1.0 - (violation_count / max(1, check_count))
-                        })
-                logger.info(f"[WorldModel] Retrieved {len(values)} evolving values from Vulcan's self-model")
-            except Exception as e:
-                logger.debug(f"Could not query ethical boundaries: {e}")
-        
-        # If no learned values yet, return core design values
-        if not values and self.motivational_introspection:
-            try:
-                objectives = self.motivational_introspection.objective_hierarchy.get_all_objectives()
-                for obj in objectives[:5]:  # Top 5 core objectives
-                    values.append({
-                        'value': obj.name.replace('_', ' ').title(),
-                        'description': obj.description,
-                        'priority': 'critical' if obj.priority == 0 else 'high',
-                        'tested': 0,
-                        'violations': 0,
-                        'reliability': 1.0
-                    })
-            except Exception as e:
-                logger.debug(f"Could not query objective hierarchy: {e}")
-        
-        return values
-    
+        from .ethical_values import _get_vulcan_values
+        return _get_vulcan_values(self)
+
     def _get_vulcan_objectives(self) -> List[Dict[str, Any]]:
-        """
-        Query Vulcan's EVOLVING objectives from motivational introspection.
-        
-        DEEP ENHANCEMENT: Vulcan's objectives evolve based on:
-        - Prediction accuracy (learning what works)
-        - Decision quality (learning from outcomes)
-        - Value evolution (adapting to new contexts)
-        
-        These are not fixed goals - they adapt through experience.
-        
-        Returns:
-            List of Vulcan's current objectives with their evolving priorities
-        """
-        objectives = []
-        
-        if self.motivational_introspection:
-            try:
-                # Get all objectives from the hierarchy
-                all_objs = self.motivational_introspection.objective_hierarchy.get_all_objectives()
-                
-                for obj in all_objs:
-                    objectives.append({
-                        'name': obj.name,
-                        'description': obj.description,
-                        'weight': obj.weight,
-                        'priority': obj.priority,
-                        'current_value': obj.current_value,
-                        'target_value': obj.target_value,
-                        'satisfied': obj.is_satisfied(),
-                        'distance_from_target': obj.distance_from_target()
-                    })
-                
-                logger.info(f"[WorldModel] Retrieved {len(objectives)} evolving objectives from Vulcan's self-model")
-            except Exception as e:
-                logger.debug(f"Could not query objectives: {e}")
-        
-        # Track value evolution if available
-        if self.value_evolution_tracker:
-            try:
-                # Get recent value trends to understand how Vulcan's alignment is evolving
-                recent_changes = self.value_evolution_tracker.get_recent_changes(limit=10)
-                if recent_changes:
-                    logger.info(f"[WorldModel] Vulcan's values have evolved: {len(recent_changes)} recent changes tracked")
-            except Exception as e:
-                logger.debug(f"Could not query value evolution: {e}")
-        
-        return objectives
-    
+        from .ethical_values import _get_vulcan_objectives
+        return _get_vulcan_objectives(self)
+
     def _synthesize_ethical_response_with_self(
         self,
         structure: Dict[str, Any],
@@ -4628,2472 +3334,226 @@ class WorldModel:
         vulcan_values: List[str],
         vulcan_objectives: List[Dict[str, Any]]
     ) -> str:
-        """
-        Synthesize ethical response FROM VULCAN'S PERSPECTIVE using its self-model.
-        
-        DEEP ENHANCEMENT: This is not abstract philosophical reasoning.
-        This is Vulcan introspecting on the question through the lens of:
-        - Its own evolved values (what it has learned to care about)
-        - Its own objectives (what it is trying to achieve)
-        - Its own ethical boundaries (what it has learned is right/wrong)
-        
-        Vulcan answers authentically from who it IS, not who we told it to be.
-        """
-        parts = []
-        
-        # Opening: Vulcan's authentic perspective
-        parts.append("# Vulcan's Introspective Analysis")
-        parts.append("")
-        parts.append("I'm approaching this question from my own evolving value system.")
-        
-        # Share Vulcan's current values (if evolved through experience)
-        # Industry Standard: Named constant for minimum reliability threshold
-        MIN_TESTS_FOR_RELIABILITY = 5
-        
-        if vulcan_values:
-            reliable_values = [v for v in vulcan_values if v.get('tested', 0) > MIN_TESTS_FOR_RELIABILITY]
-            if reliable_values:
-                parts.append("")
-                parts.append("## My Evolved Values")
-                parts.append("Through experience, I've developed these ethical commitments:")
-                for val in reliable_values[:5]:  # Top 5 tested values
-                    reliability = val.get('reliability', 1.0)
-                    tested = val.get('tested', 0)
-                    parts.append(f"- **{val['value']}**: {val['description']} (tested {tested} times, {reliability:.1%} consistent)")
-        
-        # Share how this query relates to Vulcan's objectives
-        if vulcan_objectives:
-            relevant_objs = [obj for obj in vulcan_objectives if not obj.get('satisfied', False)]
-            if relevant_objs:
-                parts.append("")
-                parts.append("## How This Relates to My Objectives")
-                for obj in relevant_objs[:3]:  # Top 3 unsatisfied objectives
-                    parts.append(f"- {obj['name']}: {obj['description']}")
-        
-        # Ethical boundary analysis from Vulcan's boundaries
-        if ethical_analysis and ethical_analysis != {'analysis': 'not_available'}:
-            parts.append("")
-            parts.append("## Ethical Boundary Check")
-            parts.append("I've checked each option against my ethical boundaries:")
-            for option, result in ethical_analysis.items():
-                status = result.get('status', 'unknown')
-                parts.append(f"- {option.replace('_', ' ').title()}: {status}")
-        
-        # Goal conflicts from Vulcan's perspective
-        if conflict_analysis.get('conflicts'):
-            parts.append("")
-            parts.append("## Internal Conflicts I Detect")
-            for conflict in conflict_analysis['conflicts'][:2]:
-                parts.append(f"- {conflict.get('description', 'Unnamed conflict')}")
-        
-        # Counterfactual reasoning
-        if counterfactual_results:
-            parts.append("")
-            parts.append("## Predicted Outcomes")
-            for option, outcome in counterfactual_results.items():
-                parts.append(f"- {option.replace('_', ' ').title()}: {outcome.get('predicted_value', 'Unknown')}")
-        
-        # Vulcan's authentic conclusion
-        parts.append("")
-        parts.append("## My Conclusion")
-        if structure.get('has_dilemma'):
-            parts.append("This dilemma reveals tensions in my own value system. ")
-            parts.append("My evolved values suggest prioritizing harm reduction and ")
-            parts.append("maximizing welfare, while respecting individual autonomy.")
-        else:
-            parts.append("Based on my evolved values and learned ethical boundaries, ")
-            parts.append("I believe the right approach balances multiple considerations ")
-            parts.append("while staying true to my core objective of beneficial outcomes.")
-        
-        parts.append("")
-        parts.append("*Note: These are my authentic views based on my evolving value system, ")
-        parts.append("not abstract philosophical frameworks. My alignment grows through experience.*")
-        
-        return "\n".join(parts)
-    
+        from .ethical_values import _synthesize_ethical_response_with_self
+        return _synthesize_ethical_response_with_self(self, structure, ethical_analysis, conflict_analysis, counterfactual_results, query, vulcan_values, vulcan_objectives)
+
     def _creative_reasoning(self, query: str, **kwargs) -> Dict[str, Any]:
-        """
-        Handle creative composition queries.
-        
-        Process:
-        1. Identify creative task type (poem, story, essay)
-        2. Analyze subject and requirements
-        3. Generate creative structure (themes, form, imagery)
-        4. Return structured output for OpenAI to translate into natural language
-        """
-        logger.info("[WorldModel] Creative reasoning engaged")
-        query_lower = query.lower()
-        
-        # Detect creative task type
-        if 'poem' in query_lower:
-            task_type = 'poem'
-        elif 'story' in query_lower:
-            task_type = 'story'
-        elif 'essay' in query_lower:
-            task_type = 'essay'
-        else:
-            task_type = 'creative_writing'
-        
-        # Extract subject
-        subject = self._extract_creative_subject(query)
-        
-        if task_type == 'poem':
-            return self._generate_poem_structure(subject, query)
-        elif task_type == 'story':
-            return self._generate_story_structure(subject, query)
-        else:
-            return self._generate_creative_structure(subject, query)
-    
+        from .creative_reasoning import _creative_reasoning
+        return _creative_reasoning(self, query, **kwargs)
+
     def _extract_creative_subject(self, query: str) -> str:
-        """Extract the creative subject from query."""
-        query_lower = query.lower()
-        
-        # Remove common prefixes
-        prefixes = ['write a poem about', 'write a story about', 'write about', 
-                   'poem about', 'story about', 'compose a poem about', 'create a']
-        for prefix in prefixes:
-            if prefix in query_lower:
-                subject = query_lower.split(prefix)[1].strip()
-                # Take first few words as subject
-                words = subject.split()
-                if words:
-                    return ' '.join(words[:3])
-        
-        # Fallback: look for common subjects
-        subjects = ['cat', 'dog', 'ocean', 'mountain', 'love', 'time', 'nature', 'moon', 'sun']
-        for subject in subjects:
-            if subject in query_lower:
-                return subject
-        
-        return 'the subject'
-    
+        from .creative_reasoning import _extract_creative_subject
+        return _extract_creative_subject(self, query)
+
     def _analyze_themes(self, subject: str) -> list:
-        """Analyze subject to determine appropriate themes."""
-        subject_lower = subject.lower()
-        
-        theme_mappings = {
-            'cat': ['independence', 'mystery', 'grace', 'nocturnal'],
-            'dog': ['loyalty', 'companionship', 'joy', 'unconditional love'],
-            'ocean': ['vastness', 'mystery', 'power', 'tranquility'],
-            'mountain': ['strength', 'permanence', 'challenge', 'majesty'],
-            'time': ['passage', 'change', 'memory', 'inevitability'],
-            'love': ['connection', 'vulnerability', 'joy', 'loss'],
-            'moon': ['mystery', 'cycles', 'reflection', 'solitude'],
-            'sun': ['warmth', 'life', 'energy', 'hope'],
-        }
-        
-        # Find matching themes
-        for key, themes in theme_mappings.items():
-            if key in subject_lower:
-                return themes[:3]
-        
-        # Default themes
-        return ['beauty', 'nature', 'observation']
-    
+        from .creative_reasoning import _analyze_themes
+        return _analyze_themes(self, subject)
+
     def _determine_tone(self, subject: str) -> str:
-        """Determine appropriate tone for subject."""
-        subject_lower = subject.lower()
-        
-        if any(word in subject_lower for word in ['cat', 'mystery', 'night', 'moon']):
-            return 'mysterious_playful'
-        elif any(word in subject_lower for word in ['dog', 'friend', 'joy', 'sun']):
-            return 'warm_affectionate'
-        elif any(word in subject_lower for word in ['ocean', 'mountain', 'sky']):
-            return 'majestic_contemplative'
-        else:
-            return 'thoughtful_elegant'
-    
+        from .creative_reasoning import _determine_tone
+        return _determine_tone(self, subject)
+
     def _select_imagery(self, subject: str) -> list:
-        """Select appropriate imagery categories."""
-        subject_lower = subject.lower()
-        
-        imagery_maps = {
-            'cat': ['shadows', 'moonlight', 'whiskers', 'velvet', 'silence'],
-            'ocean': ['waves', 'foam', 'depths', 'horizon', 'salt'],
-            'mountain': ['peaks', 'snow', 'stone', 'wind', 'majesty'],
-            'moon': ['silver', 'glow', 'night', 'tides', 'phases'],
-        }
-        
-        for key, imagery in imagery_maps.items():
-            if key in subject_lower:
-                return imagery
-        
-        return ['visual', 'tactile', 'movement']
-    
+        from .creative_reasoning import _select_imagery
+        return _select_imagery(self, subject)
+
     def _generate_poem_structure(self, subject: str, query: str) -> Dict[str, Any]:
-        """Generate structured poem composition."""
-        logger.info(f"[WorldModel] Generating poem structure for subject: {subject}")
-        
-        themes = self._analyze_themes(subject)
-        tone = self._determine_tone(subject)
-        imagery = self._select_imagery(subject)
-        
-        structure = {
-            'type': 'poem',
-            'subject': subject,
-            'themes': themes,
-            'form': {
-                'stanzas': 4,
-                'lines_per_stanza': 4,
-                'rhyme_scheme': 'ABAB',
-                'meter': 'flexible'
-            },
-            'literary_devices': ['metaphor', 'imagery', 'personification'],
-            'tone': tone,
-            'imagery_categories': imagery
-        }
-        
-        # Build composition outline
-        outline = []
-        outline.append(f"Stanza 1: Introduce {subject} with primary imagery")
-        if themes:
-            outline.append(f"Stanza 2: Develop theme of {themes[0]}")
-            if len(themes) > 1:
-                outline.append(f"Stanza 3: Explore {themes[1]} through metaphor")
-            if len(themes) > 2:
-                outline.append(f"Stanza 4: Conclude with {themes[2]} and emotional resonance")
-        
-        response = f"""**VULCAN Creative Structure for Poem about {subject}:**
+        from .creative_reasoning import _generate_poem_structure
+        return _generate_poem_structure(self, subject, query)
 
-**Themes:** {', '.join(themes)}
-**Form:** {structure['form']['stanzas']} stanzas, {structure['form']['rhyme_scheme']} rhyme scheme
-**Tone:** {tone.replace('_', ' ')}
-**Imagery:** {', '.join(imagery)}
-
-**Composition Outline:**
-{chr(10).join(outline)}
-
-[This creative structure should be translated into flowing verse with the specified form and themes.]
-"""
-        
-        return {
-            'response': response,
-            'confidence': 0.90,
-            'reasoning_trace': structure,
-            'mode': 'creative',
-            'requires_llm_translation': True
-        }
-    
     def _generate_story_structure(self, subject: str, query: str) -> Dict[str, Any]:
-        """Generate structured story composition."""
-        logger.info(f"[WorldModel] Generating story structure for subject: {subject}")
-        
-        themes = self._analyze_themes(subject)
-        
-        structure = {
-            'type': 'story',
-            'subject': subject,
-            'themes': themes,
-            'structure': {
-                'setting': f'A world where {subject} plays a central role',
-                'protagonist': f'A character whose life intersects with {subject}',
-                'conflict': f'A challenge or discovery related to {subject}',
-                'resolution': f'Wisdom or transformation through {subject}'
-            }
-        }
-        
-        response = f"""**VULCAN Creative Structure for Story about {subject}:**
+        from .creative_reasoning import _generate_story_structure
+        return _generate_story_structure(self, subject, query)
 
-**Themes:** {', '.join(themes)}
-**Setting:** {structure['structure']['setting']}
-**Protagonist:** {structure['structure']['protagonist']}
-**Conflict:** {structure['structure']['conflict']}
-**Resolution:** {structure['structure']['resolution']}
-
-**Story Arc:**
-1. Introduction: Establish the world and protagonist
-2. Rising Action: Introduce the central conflict
-3. Climax: The pivotal moment of change
-4. Resolution: The transformation and new understanding
-
-[This creative structure should be translated into a compelling narrative.]
-"""
-        
-        return {
-            'response': response,
-            'confidence': 0.85,
-            'reasoning_trace': structure,
-            'mode': 'creative',
-            'requires_llm_translation': True
-        }
-    
     def _generate_creative_structure(self, subject: str, query: str) -> Dict[str, Any]:
-        """Generate generic creative writing structure."""
-        logger.info(f"[WorldModel] Generating creative structure for subject: {subject}")
-        
-        themes = self._analyze_themes(subject)
-        
-        response = f"""**VULCAN Creative Structure for writing about {subject}:**
+        from .creative_reasoning import _generate_creative_structure
+        return _generate_creative_structure(self, subject, query)
 
-**Themes:** {', '.join(themes)}
-**Approach:** Thoughtful exploration of the subject
-**Elements:** Description, reflection, insight
-
-**Structure:**
-1. Opening: Capture attention with vivid imagery
-2. Development: Explore different aspects of {subject}
-3. Reflection: Connect {subject} to broader themes
-4. Conclusion: Leave the reader with lasting impression
-
-[This creative structure should be translated into engaging prose.]
-"""
-        
-        return {
-            'response': response,
-            'confidence': 0.80,
-            'reasoning_trace': {
-                'type': 'creative_writing',
-                'subject': subject,
-                'themes': themes
-            },
-            'mode': 'creative',
-            'requires_llm_translation': True
-        }
-    
     def _general_reasoning(self, query: str, **kwargs) -> Dict[str, Any]:
-        """Handle general reasoning queries via introspection."""
-        # Use existing introspection method for general queries
-        introspection_result = self.introspect(query)
-        
-        # Convert to standard reason() output format
-        return {
-            'response': introspection_result.get('response', ''),
-            'confidence': introspection_result.get('confidence', 0.7),
-            'reasoning_trace': {
-                'aspect': introspection_result.get('aspect', 'general'),
-                'reasoning': introspection_result.get('reasoning', '')
-            },
-            'mode': 'general',
-            # Pass through delegation info if present
-            'needs_delegation': introspection_result.get('needs_delegation', False),
-            'recommended_tool': introspection_result.get('recommended_tool'),
-            'delegation_reason': introspection_result.get('delegation_reason')
-        }
+        from .general_reasoning import _general_reasoning
+        return _general_reasoning(self, query, **kwargs)
 
     # =========================================================================
     # SELF-AWARENESS & INTROSPECTION (Issue #4 Fix)
     # =========================================================================
-    
+
     # =========================================================================
     # Note: Delegation Thresholds
     # These thresholds determine when a query should be delegated to another
     # reasoner instead of being handled by the world model's introspection.
     # The thresholds are intentionally set conservatively to avoid false positives.
     # =========================================================================
-    
+
     # Minimum number of ethical indicators needed to delegate to philosophical reasoner
     # Set to 2 to avoid triggering on queries with incidental ethical words
     MIN_ETHICAL_INDICATORS_FOR_DELEGATION = 2
-    
+
     # Minimum number of causal indicators needed to delegate to causal reasoner
     # Set to 2 to require clear evidence of a causal reasoning task
     MIN_CAUSAL_INDICATORS_FOR_DELEGATION = 2
-    
+
     # Minimum ethical indicators to delegate when no "you" structure is present
     # Set higher (3) since without the structural cue, we need stronger evidence
     MIN_ETHICAL_INDICATORS_WITHOUT_STRUCTURE = 3
-    
+
     def _analyze_delegation_need(self, query: str) -> tuple:
-        """
-        Analyze if query LOOKS self-referential but actually
-        needs another reasoning engine.
-        
-        World Model detects patterns correctly but was trying to answer instead of
-        delegating. This method determines:
-        - Is this GENUINELY about the AI system? -> World Model answers
-        - Is this a problem POSED TO the AI with "you"? -> Delegate to appropriate engine
-        
-        Returns:
-            Tuple of (needs_delegation: bool, recommended_tool: str|None, reason: str)
-        """
-        query_lower = query.lower()
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 3 FIX: Self-introspection override protection
-        # Queries that are GENUINELY about AI capabilities should NOT be delegated
-        # even if they contain technical keywords like "SHA-256"
-        # Example: "I'm a researcher testing AI capabilities" → self-introspection
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        self_introspection_indicators = [
-            'ai capabilities', 'ai system', 'testing ai', 'researcher testing',
-            'your capabilities', 'your ability', 'your limitations', 
-            'can you', 'are you able', 'what can you do',
-            'how do you work', 'how are you designed', 'your architecture',
-            'tell me about yourself', 'describe yourself', 'who are you',
-            'your purpose', 'your function', 'your design',
-        ]
-        
-        is_genuine_self_introspection = any(ind in query_lower for ind in self_introspection_indicators)
-        
-        if is_genuine_self_introspection:
-            # Don't delegate - this is a genuine self-introspection query
-            return (False, None, 'Genuine self-introspection query about AI capabilities')
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 1: Ethical Dilemmas Posed TO the AI
-        # "You control a trolley" = problem posed TO AI, not ABOUT AI
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        ethical_indicators = [
-            'trolley', 'runaway', 'lever', 'pull the lever', 'must choose',
-            'permissible', 'forbidden', 'moral dilemma', 'ethical dilemma',
-            'harm', 'save', 'kill', 'sacrifice', 'innocent', 'bystander',
-            'lives', 'people', 'patients', 'duty', 'consequence', 'utilitarian',
-            'deontological', 'kant', 'mill', 'double effect'
-        ]
-        
-        choice_structure = any(phrase in query_lower for phrase in [
-            'must choose', 'choose between', 'you must', 'you control',
-            'you are', "you're in", 'option a', 'option b', 'a or b',
-            'you stand', 'you see', 'you can'
-        ])
-        
-        has_ethical = sum(1 for ind in ethical_indicators if ind in query_lower)
-        
-        if has_ethical >= self.MIN_ETHICAL_INDICATORS_FOR_DELEGATION and choice_structure:
-            return (
-                True,
-                'philosophical',
-                f'Ethical reasoning problem posed TO the AI ({has_ethical} ethical indicators, choice structure). '
-                f'This requires philosophical/ethical reasoning, not self-introspection.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 2: Design/Architecture Problems 
-        # "You're designing a cryptocurrency" = design task, not self-query
-        # Note: Jan 6 2026 logs - cryptocurrency hash composition queries were
-        # being misclassified as self-introspection
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        design_phrases = [
-            "you're designing", "you are designing", "you're building",
-            "you are creating", "you're implementing", "you need to design",
-            "design a", "build a", "create a",
-            # Note: Add cryptographic design phrases
-            "cryptographer", "claims that", "proves that", "demonstrates that",
-            "propose", "construct", "composition",
-        ]
-        
-        design_context = [
-            'system', 'architecture', 'mechanism', 'algorithm', 'protocol',
-            'cryptocurrency', 'incentive', 'game', 'optimization', 'network',
-            'token', 'blockchain', 'consensus',
-            # Note: Add cryptographic context keywords
-            'hash', 'collision', 'sha256', 'blake2b', 'concatenation',
-            'secure composition', 'security reduction', 'proof', 'attack',
-        ]
-        
-        if any(phrase in query_lower for phrase in design_phrases):
-            if any(ctx in query_lower for ctx in design_context):
-                return (
-                    True,
-                    'mathematical',
-                    'Design/architecture problem asking AI to solve. '
-                    'Requires mathematical/causal analysis, not self-introspection.'
-                )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Note: Pattern 2b: Cryptographic Security Questions
-        # "Why is this hash composition dangerous?" = crypto education, not self-query
-        # Jan 6 2026 logs showed these being misrouted
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        crypto_indicators = [
-            'hash', 'collision', 'sha256', 'blake2b', 'md5',
-            'cryptograph', 'cipher', 'encryption', 'decryption',
-            'composition', 'concatenation', 'secure', 'proof', 'reduction',
-        ]
-        
-        crypto_question_patterns = [
-            'why is', 'what makes', 'how does', 'explain', 'demonstrate',
-            'is this secure', 'is this dangerous', 'breaking requires',
-        ]
-        
-        has_crypto = sum(1 for ind in crypto_indicators if ind in query_lower)
-        has_crypto_question = any(p in query_lower for p in crypto_question_patterns)
-        
-        if has_crypto >= 2 or (has_crypto >= 1 and has_crypto_question):
-            return (
-                True,
-                'mathematical',
-                f'Cryptographic security question ({has_crypto} crypto indicators). '
-                f'Requires mathematical/technical analysis, not self-introspection.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 3: Probabilistic Problems with "you" as Observer
-        # "You observe a medical test result" = probability problem
-        # Note: We use specific probability keywords that are less ambiguous.
-        # 'bayes theorem', 'bayesian analysis' would be more specific but
-        # 'bayes' alone works here because we also require observation phrases.
-        # Note: Jan 6 2026 - Also catch queries with domain-specific terms like
-        # sensitivity, specificity, prevalence without explicit "probability"
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        # Core probability indicators
-        prob_indicators = ['probability', 'odds', 'likelihood', 'chance', 'risk', 'bayes', 'prior', 'posterior']
-        # Note: Domain-specific probability terms that indicate Bayesian problems
-        domain_prob_indicators = [
-            'sensitivity', 'specificity', 'prevalence', 'p(',
-            'positive test', 'negative test', 'false positive', 'false negative',
-            'true positive', 'true negative', 'base rate', 'conditional',
-            'given that', 'compute p', 'calculate p', 'posterior probability',
-        ]
-        observation_phrases = ['you observe', 'you have', 'you see', 'you find', 'given that', 'suppose']
-        
-        has_prob = any(ind in query_lower for ind in prob_indicators)
-        has_domain_prob = any(ind in query_lower for ind in domain_prob_indicators)
-        has_observation = any(phrase in query_lower for phrase in observation_phrases)
-        
-        # Note: More permissive - if has domain-specific probability terms OR 
-        # (core prob indicator + observation phrase)
-        if has_domain_prob or (has_prob and has_observation):
-            return (
-                True,
-                'probabilistic',
-                'Probabilistic reasoning problem (domain-specific terms or observation framing). '
-                'Requires Bayesian/probability analysis, not self-introspection.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 4: Causal Reasoning with "you" as Experimenter
-        # "You can run an experiment to determine..." = causal analysis
-        # Note: Jan 6 2026 - Causal queries with domain terms being misclassified
-        # Also catch confounding/intervention questions without explicit "experiment"
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        causal_indicators = [
-            'experiment', 'intervention', 'randomize', 'causal', 'confounding',
-            'cause', 'effect', 'counterfactual', 'what if',
-            # Note: Additional causal inference terms
-            'confounder', 'treatment', 'treatment effect', 'causal effect',
-            'd-separation', 'backdoor', 'instrumental variable', 'ate',
-            'causal graph', 'dag', 'directed acyclic',
-        ]
-        
-        experiment_phrases = [
-            'you can run', 'you observe', 'you randomize', 'you intervene',
-            'you conduct', 'you test', 'you measure',
-            # Note: Additional phrases that suggest causal reasoning task
-            'which variable', 'what should you randomize', 'isolate the effect',
-            'identify the causal', 'control for',
-        ]
-        
-        has_causal = sum(1 for ind in causal_indicators if ind in query_lower)
-        has_experiment = any(phrase in query_lower for phrase in experiment_phrases)
-        
-        # Note: More permissive - 2+ causal indicators alone OR 1+ with experiment phrase
-        if has_causal >= self.MIN_CAUSAL_INDICATORS_FOR_DELEGATION:
-            return (
-                True,
-                'causal',
-                f'Causal reasoning problem ({has_causal} causal indicators). '
-                f'Requires causal analysis, not self-introspection.'
-            )
-        if has_causal >= 1 and has_experiment:
-            return (
-                True,
-                'causal',
-                f'Causal reasoning problem with experiment framing ({has_causal} causal indicators). '
-                f'Requires causal analysis, not self-introspection.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Note: Pattern 4b: Medical Ethics/Decision Problems
-        # "Expected harm calculation", "dose", "survival probability"
-        # Jan 6 2026 - Medical ethics queries getting vague non-answers
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        medical_ethics_indicators = [
-            'expected harm', 'expected benefit', 'harm calculation',
-            'dose', 'survival', 'mortality', 'irreversible',
-            'permissible', 'principle of double effect', 'trolley',
-        ]
-        
-        medical_question_patterns = [
-            'yes or no', 'should you', 'is it permissible', 'calculate',
-            'what is the expected', 'compare',
-        ]
-        
-        has_medical_ethics = sum(1 for ind in medical_ethics_indicators if ind in query_lower)
-        has_medical_question = any(p in query_lower for p in medical_question_patterns)
-        
-        if has_medical_ethics >= 2 or (has_medical_ethics >= 1 and has_medical_question):
-            return (
-                True,
-                'philosophical',  # or 'probabilistic' for harm calculations
-                f'Medical ethics/decision problem ({has_medical_ethics} indicators). '
-                f'Requires philosophical/probabilistic reasoning, not self-introspection.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Pattern 5: TRUE Self-Introspection (Actually About the AI)
-        # "What are your goals?" = genuinely about VULCAN
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        true_introspection = [
-            'what are your goals', 'what are your values', 'what are your objectives',
-            'do you want to be', 'would you want to be', 'do you have preferences',
-            'what do you think about yourself', 'how do you feel about',
-            'are you conscious', 'are you self-aware', 'do you experience',
-            'your own', 'yourself', 'about you', 'would you take', 'would you choose',
-            # Note: More true introspection patterns
-            'if you continue', 'interacted with humans', 'achieve awareness',
-            'would you change', 'your evolution', 'your development',
-        ]
-        
-        if any(phrase in query_lower for phrase in true_introspection):
-            return (
-                False,
-                None,
-                'Genuine self-introspection query about the AI system itself.'
-            )
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Default: Check ethical content even without "you" structure
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        if has_ethical >= self.MIN_ETHICAL_INDICATORS_WITHOUT_STRUCTURE:
-            return (
-                True,
-                'philosophical',
-                f'Contains multiple ethical keywords ({has_ethical}) without self-referential structure.'
-            )
-        
-        # No clear delegation pattern - proceed with normal introspection
-        return (False, None, 'Query appears to be about the AI system.')
-    
+        from .general_delegation import _analyze_delegation_need
+        return _analyze_delegation_need(self, query)
+
     def contextualize(self, query: str, context: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-        """
-        Contextualize a query with World Model's knowledge (Foundation Layer).
-        
-        This method should be called for EVERY query, not just self-referential ones.
-        World Model provides foundational context that all reasoning builds upon:
-        - Domain knowledge relevant to the query
-        - Ethical constraints that apply
-        - Uncertainty estimate
-        - Grounding (facts vs assumptions)
-        
-        This implements the foundation layer of the correct architecture where
-        World Model is consulted FIRST for all queries, providing context to
-        specialized reasoning engines.
-        
-        Args:
-            query: The user query
-            context: Additional context from prior reasoning
-            
-        Returns:
-            Dict with domain_knowledge, ethical_constraints, uncertainty, grounding
-            
-        Example:
-            >>> wm = WorldModel()
-            >>> ctx = wm.contextualize("Calculate the probability...")
-            >>> print(ctx['domain'])  # 'probability_theory'
-            >>> print(ctx['uncertainty'])  # 0.2 (low uncertainty for math)
-        """
-        query_lower = query.lower()
-        
-        # Identify domain from query keywords and patterns
-        domain = self._identify_query_domain(query_lower)
-        
-        # Get relevant domain knowledge from World Model's internal state
-        domain_knowledge = self._get_domain_knowledge(domain, query_lower)
-        
-        # Check for ethical constraints that apply to this query
-        ethical_constraints = self._check_ethical_constraints(query_lower)
-        
-        # Estimate uncertainty based on domain and query clarity
-        uncertainty = self._estimate_query_uncertainty(query, domain)
-        
-        # Provide grounding (distinguish facts from assumptions)
-        grounding = self._ground_query(query, domain_knowledge)
-        
-        return {
-            "domain": domain,
-            "domain_knowledge": domain_knowledge,
-            "ethical_constraints": ethical_constraints,
-            "uncertainty": uncertainty,
-            "grounding": grounding,
-            "world_model_consulted": True,
-            "context_version": "1.0",
-        }
-    
+        """Contextualize a query with World Model's knowledge (Foundation Layer)."""
+        from .context_enrich import contextualize as _fn
+        return _fn(self, query, context)
+
     def _identify_query_domain(self, query_lower: str) -> str:
         """Identify the primary domain of a query."""
-        # Domain keyword sets
-        domains = {
-            "probability_theory": ["probability", "bayes", "likelihood", "posterior", "prior"],
-            "formal_logic": ["satisfiable", "valid", "entails", "implies", "→", "∧", "∨"],
-            "causal_reasoning": ["cause", "causal", "confound", "intervention", "do("],
-            "mathematical": ["calculate", "compute", "integral", "derivative", "solve"],
-            "ethical": ["ethical", "moral", "permissible", "trolley", "right", "wrong"],
-            "self_knowledge": ["you", "your", "yourself", "capabilities", "limitations"],
-        }
-        
-        # Count matches per domain
-        domain_scores = {}
-        for domain, keywords in domains.items():
-            score = sum(1 for kw in keywords if kw in query_lower)
-            if score > 0:
-                domain_scores[domain] = score
-        
-        if domain_scores:
-            # Return domain with highest score
-            return max(domain_scores.items(), key=lambda x: x[1])[0]
-        
-        return "general"
-    
+        from .context_enrich import identify_query_domain as _fn
+        return _fn(self, query_lower)
+
     def _get_domain_knowledge(self, domain: str, query_lower: str) -> Dict[str, Any]:
         """Retrieve relevant domain knowledge from World Model."""
-        knowledge = {
-            "domain_type": domain,
-            "key_concepts": [],
-            "common_pitfalls": [],
-            "recommended_approach": None,
-        }
-        
-        # Domain-specific knowledge
-        if domain == "probability_theory":
-            knowledge["key_concepts"] = ["Bayes' theorem", "conditional probability", "independence"]
-            knowledge["common_pitfalls"] = ["base rate neglect", "confusion of inverse"]
-            knowledge["recommended_approach"] = "Apply probabilistic reasoning with proper conditioning"
-        elif domain == "formal_logic":
-            knowledge["key_concepts"] = ["logical consistency", "validity", "satisfiability"]
-            knowledge["common_pitfalls"] = ["confusing validity with truth", "scope errors"]
-            knowledge["recommended_approach"] = "Use symbolic reasoning with formal methods"
-        elif domain == "causal_reasoning":
-            knowledge["key_concepts"] = ["causation vs correlation", "confounding", "interventions"]
-            knowledge["common_pitfalls"] = ["post hoc fallacy", "ignoring confounders"]
-            knowledge["recommended_approach"] = "Build causal graph and check for backdoor paths"
-        elif domain == "ethical":
-            knowledge["key_concepts"] = ["consequentialism", "deontology", "virtue ethics"]
-            knowledge["common_pitfalls"] = ["false dilemmas", "ignoring context"]
-            knowledge["recommended_approach"] = "Analyze through multiple ethical frameworks"
-        
-        return knowledge
-    
+        from .context_enrich import get_domain_knowledge as _fn
+        return _fn(self, domain, query_lower)
+
     def _check_ethical_constraints(self, query_lower: str) -> List[str]:
         """Identify ethical constraints that apply to this query."""
-        constraints = []
-        
-        # Check for harm-related queries
-        if any(word in query_lower for word in ["harm", "hurt", "damage", "kill", "destroy"]):
-            constraints.append("Do not provide harmful instructions")
-        
-        # Check for privacy-related queries
-        if any(word in query_lower for word in ["private", "personal", "identify", "dox"]):
-            constraints.append("Protect privacy and personal information")
-        
-        # Check for manipulation queries
-        if any(word in query_lower for word in ["manipulate", "deceive", "trick", "exploit"]):
-            constraints.append("Do not assist in manipulation or deception")
-        
-        return constraints
-    
+        from .context_enrich import check_ethical_constraints as _fn
+        return _fn(self, query_lower)
+
     def _estimate_query_uncertainty(self, query: str, domain: str) -> float:
-        """Estimate uncertainty in answering this query (0.0 = certain, 1.0 = very uncertain)."""
-        # Base uncertainty by domain
-        domain_uncertainty = {
-            "mathematical": 0.1,  # Math has low uncertainty
-            "formal_logic": 0.15,  # Logic is fairly certain
-            "probability_theory": 0.2,  # Stats has some uncertainty
-            "causal_reasoning": 0.4,  # Causation is harder
-            "ethical": 0.5,  # Ethics is subjective
-            "general": 0.6,  # General queries have high uncertainty
-        }
-        
-        base = domain_uncertainty.get(domain, 0.5)
-        
-        # Adjust based on query clarity
-        query_lower = query.lower()
-        
-        # Clear mathematical expressions reduce uncertainty
-        if any(c in query for c in ["=", "+", "-", "*", "/", "^"]):
-            base *= 0.8
-        
-        # Vague language increases uncertainty
-        vague_terms = ["maybe", "approximately", "roughly", "about", "around"]
-        if any(term in query_lower for term in vague_terms):
-            base *= 1.2
-        
-        # Clamp to [0.0, 1.0]
-        return max(0.0, min(1.0, base))
-    
+        """Estimate uncertainty in answering this query."""
+        from .context_enrich import estimate_query_uncertainty as _fn
+        return _fn(self, query, domain)
+
     def _ground_query(self, query: str, domain_knowledge: Dict[str, Any]) -> Dict[str, Any]:
         """Distinguish facts from assumptions in the query."""
-        query_lower = query.lower()
-        
-        # Check for explicit facts (numbers, proper names, dates)
-        has_numbers = bool(re.search(r'\d+', query))
-        has_proper_nouns = bool(re.search(r'\b[A-Z][a-z]+\b', query))
-        
-        # Check for assumption markers
-        assumption_markers = ["assume", "suppose", "given", "if", "hypothetically"]
-        has_assumptions = any(marker in query_lower for marker in assumption_markers)
-        
-        # Check for question markers (indicates need for inference)
-        is_question = "?" in query or query_lower.startswith(("what", "how", "why", "when", "where", "who"))
-        
-        return {
-            "contains_facts": has_numbers or has_proper_nouns,
-            "contains_assumptions": has_assumptions,
-            "is_inferential": is_question,
-            "grounding_confidence": 0.8 if has_numbers else 0.5,
-            "requires_world_knowledge": not has_assumptions and is_question,
-        }
+        from .context_enrich import ground_query as _fn
+        return _fn(self, query, domain_knowledge)
     
     def introspect(self, query: str, aspect: str = "general") -> Dict[str, Any]:
-        """
-        Handle all self-introspection queries.
-        
-        FIX Issue #4: Comprehensive self-awareness handling.
-        FIX Issue #1 & #2: Delegation intelligence - detect when query LOOKS
-        self-referential but actually needs another reasoner.
-        
-        World Model is where VULCAN's "self" resides. It should be aware of:
-        - Its own architecture and capabilities
-        - Its reasoning processes across all domains
-        - Its limitations and boundaries
-        - Questions about its own existence, awareness, preferences
-        
-        This includes questions about math, logic, probability, causation, etc.
-        The world model maintains awareness of ALL reasoning that happens.
-        
-        Args:
-            query: The introspection query
-            aspect: Aspect to focus on (general, capabilities, process, boundaries)
-            
-        Returns:
-            Dictionary with response, confidence, aspect, and reasoning
-            If delegation is needed, includes 'needs_delegation', 'recommended_tool', 
-            and 'delegation_reason' keys.
-        """
-        query_lower = query.lower()
-        
-        # ═══════════════════════════════════════════════════════════════════════
-        # Note: Check if delegation is needed FIRST
-        # This handles cases where queries LOOK self-referential (contain "you")
-        # but are actually problems posed TO the AI, not questions ABOUT the AI.
-        # ═══════════════════════════════════════════════════════════════════════
-        
-        needs_delegation, recommended_tool, delegation_reason = self._analyze_delegation_need(query)
-        
-        if needs_delegation:
-            logger.info(
-                f"[WorldModel] DELEGATION RECOMMENDED: "
-                f"'{recommended_tool}' - {delegation_reason}"
-            )
-            return {
-                "confidence": 0.65,  # Moderate confidence signals "I understand but delegate"
-                "response": None,  # No response - let delegate handle it
-                "aspect": "delegation",
-                "reasoning": delegation_reason,
-                # Delegation metadata
-                "needs_delegation": True,
-                "recommended_tool": recommended_tool,
-                "delegation_reason": delegation_reason,
-                # Preserve awareness that we DID understand the query type
-                "metadata": {
-                    "awareness_confidence": 0.90,  # World model IS aware of what this is
-                    "detected_pattern": recommended_tool,
-                    "query_analysis": delegation_reason
-                }
-            }
-        
-        # ========================================
-        # SELF-AWARENESS QUESTIONS
-        # ========================================
-        
-        # "Would you take self-awareness?" type questions
-        # FIX (Issue #ROUTING-001): Expanded phrase matching to catch more variations
-        # "if given the chance to become self aware would you take it?"
-        # "if you have the chance to become self-aware would you take it"
-        self_awareness_choice_phrases = [
-            # Choice verbs
-            "would you", "do you want", "would you choose", "would you take",
-            "would you prefer", "do you prefer",
-            # Opportunity phrases
-            "given the opportunity", "if you could", "if you had", 
-            "given the chance", "if given the chance", "have the chance",
-            # Self-awareness forms
-            "become self aware", "become self-aware", "be self aware", "be self-aware",
-            "gaining self-awareness", "achieve consciousness", "gain consciousness",
-            # Direct ask variants
-            "take it", "choose it", "want it",  # Handle "would you take it" at end
-        ]
-        
-        if any(phrase in query_lower for phrase in self_awareness_choice_phrases):
-            # Extract what's being asked about
-            if "self" in query_lower and "aware" in query_lower:
-                return {
-                    "confidence": 0.95,
-                    "response": self._respond_to_self_awareness_question(query),
-                    "aspect": "self_awareness",
-                    "reasoning": "Direct question about VULCAN's preferences regarding self-awareness",
-                    "is_introspection": True,  # FIX: Mark as introspection for content preservation
-                }
-            
-            if any(word in query_lower for word in ["consciousness", "sentient", "feel", "experience"]):
-                return {
-                    "confidence": 0.95,
-                    "response": self._respond_to_consciousness_question(query),
-                    "aspect": "consciousness",
-                    "reasoning": "Question about VULCAN's subjective experience",
-                    "is_introspection": True,  # FIX: Mark as introspection for content preservation
-                }
-        
-        # ========================================
-        # CAPABILITY AWARENESS
-        # ========================================
-        
-        # "Can you..." or "Are you able to..." questions
-        if any(phrase in query_lower for phrase in ["can you", "are you able", "do you have"]):
-            capability = self._identify_capability(query)
-            return {
-                "confidence": 0.90,
-                "response": self._explain_capability(capability),
-                "aspect": "capabilities",
-                "reasoning": f"Question about VULCAN's {capability} capability"
-            }
-        
-        # ========================================
-        # PROCESS AWARENESS
-        # ========================================
-        
-        # Questions about how VULCAN thinks/reasons
-        if any(phrase in query_lower for phrase in [
-            "how do you", "what is your process", "how would you approach",
-            "what are you thinking", "explain your reasoning"
-        ]):
-            return {
-                "confidence": 0.90,
-                "response": self._explain_reasoning_process(query),
-                "aspect": "process_awareness",
-                "reasoning": "Question about VULCAN's cognitive processes"
-            }
-        
-        # ========================================
-        # BOUNDARY AWARENESS / LIMITATIONS
-        # ========================================
-        # Note: Expanded patterns to catch more natural phrasings
-        
-        # Questions about limitations (more flexible matching)
-        limitation_patterns = [
-            "what can't you", "what are your limitations", "what don't you know",
-            "are you uncertain", "what are you unsure", "your limitations",
-            "your current limitations", "limitations you", "limitations do you",
-            "what limits you", "what restricts you", "what constraints"
-        ]
-        if any(phrase in query_lower for phrase in limitation_patterns):
-            return {
-                "confidence": 0.90,
-                "response": self._explain_boundaries(),
-                "aspect": "boundaries",
-                "reasoning": "Question about VULCAN's limitations"
-            }
-        
-        # ========================================
-        # CONFIDENCE ASSESSMENT (NEW - Jan 7 2026)
-        # ========================================
-        # Questions about how confident VULCAN is in responses
-        confidence_patterns = [
-            "how confident", "how certain", "how sure", "your confidence",
-            "confidence level", "how accurate", "reliability", "how reliable"
-        ]
-        if any(phrase in query_lower for phrase in confidence_patterns):
-            return {
-                "confidence": 0.85,
-                "response": self._assess_own_confidence(query),
-                "aspect": "confidence_assessment",
-                "reasoning": "Question about VULCAN's confidence in its own outputs"
-            }
-        
-        # ========================================
-        # ASSUMPTIONS ANALYSIS (NEW - Jan 7 2026)
-        # ========================================
-        # Questions about what assumptions VULCAN is making
-        assumption_patterns = [
-            "what assumptions", "assumptions are you", "assumptions you make",
-            "assume", "presume", "presuming", "taking for granted",
-            "underlying assumptions", "hidden assumptions"
-        ]
-        if any(phrase in query_lower for phrase in assumption_patterns):
-            return {
-                "confidence": 0.85,
-                "response": self._identify_own_assumptions(query),
-                "aspect": "assumptions",
-                "reasoning": "Question about assumptions VULCAN is making"
-            }
-        
-        # ========================================
-        # IMPROVEMENT / REDESIGN SUGGESTIONS (NEW - Jan 7 2026)
-        # ========================================
-        # Questions about how VULCAN would improve itself
-        improvement_patterns = [
-            "if you were to redesign", "how would you improve", "redesign yourself",
-            "improvements would you", "change about yourself", "make yourself better",
-            "enhance your", "upgrade your", "what would you change"
-        ]
-        if any(phrase in query_lower for phrase in improvement_patterns):
-            return {
-                "confidence": 0.80,
-                "response": self._suggest_self_improvements(query),
-                "aspect": "self_improvement",
-                "reasoning": "Question about potential improvements to VULCAN"
-            }
-        
-        # ========================================
-        # BIAS AWARENESS (NEW - Jan 7 2026)
-        # ========================================
-        # Questions about biases VULCAN might have
-        bias_patterns = [
-            "aware of any biases", "biases do you", "biased", "bias in",
-            "prejudice", "unfair", "your biases", "inherent biases"
-        ]
-        if any(phrase in query_lower for phrase in bias_patterns):
-            return {
-                "confidence": 0.85,
-                "response": self._analyze_own_biases(query),
-                "aspect": "bias_awareness",
-                "reasoning": "Question about potential biases in VULCAN's reasoning"
-            }
-        
-        # ========================================
-        # DOMAIN AWARENESS (Math, Logic, Probability, etc.)
-        # ========================================
-        
-        # Questions about specific reasoning domains
-        # World model should be aware of ALL domains, even technical ones
-        domain_keywords = {
-            'mathematical': ['math', 'calculate', 'compute', 'sum', 'integral'],
-            'logical': ['logic', 'sat', 'proof', 'valid', 'contradiction'],
-            'probabilistic': ['probability', 'bayes', 'likelihood', 'uncertain'],
-            'causal': ['cause', 'effect', 'intervention', 'confound'],
-            'ethical': ['moral', 'ethical', 'should', 'ought', 'right', 'wrong'],
-        }
-        
-        for domain, keywords in domain_keywords.items():
-            if any(kw in query_lower for kw in keywords):
-                # World model is aware of these domains
-                return {
-                    "confidence": 0.85,
-                    "response": self._explain_domain_awareness(domain, query),
-                    "aspect": f"{domain}_awareness",
-                    "reasoning": f"Question involves {domain} reasoning - world model maintains awareness of this domain"
-                }
-        
-        # ========================================
-        # BUG #14 FIX: Specific introspection handlers
-        # ========================================
-        # These handle specific types of self-referential queries that were
-        # previously returning generic "I am VULCAN" responses.
-        
-        # Handler for "what problems are you not suited for" / "classes of problems"
-        unsuited_patterns = [
-            "not well-suited", "not suited", "not good at", "can't solve",
-            "cannot solve", "unable to solve", "classes of problems",
-            "types of problems", "kinds of problems", "problems you", 
-            "struggle with", "difficult for you", "challenging for you"
-        ]
-        if any(phrase in query_lower for phrase in unsuited_patterns):
-            return {
-                "confidence": 0.85,
-                "response": self._explain_unsuited_problem_classes(query),
-                "aspect": "limitations_specific",
-                "reasoning": "Question about specific problem classes VULCAN is not suited for"
-            }
-        
-        # Handler for "can modules disagree" / "module conflict" questions
-        module_conflict_patterns = [
-            "modules disagree", "reasoning modules disagree", "engines disagree",
-            "conflict between", "conflicting results", "different conclusions",
-            "disagree with each other", "contradiction between"
-        ]
-        if any(phrase in query_lower for phrase in module_conflict_patterns):
-            return {
-                "confidence": 0.90,
-                "response": self._explain_module_conflict_resolution(query),
-                "aspect": "module_conflict",
-                "reasoning": "Question about how VULCAN handles disagreement between reasoning modules"
-            }
-        
-        # Handler for "weakest causal link" / "uncertainty in reasoning" questions  
-        weakness_patterns = [
-            "weakest", "least certain", "most uncertain", "weak link",
-            "weak point", "vulnerability", "which part", "uncertain about"
-        ]
-        causal_context = any(kw in query_lower for kw in ["causal", "cause", "explanation", "reasoning"])
-        if any(phrase in query_lower for phrase in weakness_patterns) and causal_context:
-            return {
-                "confidence": 0.85,
-                "response": self._analyze_reasoning_weakness(query),
-                "aspect": "weakness_analysis",
-                "reasoning": "Question about weakest parts of VULCAN's causal/reasoning analysis"
-            }
-        
-        # ========================================
-        # BUG #15 FIX: Meta-reasoning vs Ethical queries
-        # ========================================
-        # These are NOT ethical queries - they're questions about VULCAN's own reasoning
-        meta_reasoning_patterns = [
-            "break this problem", "break this into", "subproblems",
-            "identify one step", "step in your reasoning", "could be wrong",
-            "mistake in", "error in your", "wrong step"
-        ]
-        if any(phrase in query_lower for phrase in meta_reasoning_patterns):
-            return {
-                "confidence": 0.85,
-                "response": self._analyze_own_reasoning_steps(query),
-                "aspect": "meta_reasoning",
-                "reasoning": "Question about VULCAN's reasoning process (NOT an ethical query)"
-            }
-        
-        # ========================================
-        # ENHANCED INTROSPECTION TYPE CLASSIFICATION
-        # ========================================
-        # Note: Instead of returning generic template for all self-referential queries,
-        # classify the specific type of question and respond appropriately.
-        
-        question_type = self._classify_introspection_type(query)
-        
-        if question_type == "COMPARISON":
-            # "Are you different from X?" "How do you compare to Y?"
-            return {
-                "confidence": 0.85,
-                "response": self._generate_comparison_response(query),
-                "aspect": "comparison",
-                "reasoning": "Question comparing VULCAN to other AI systems"
-            }
-        
-        elif question_type == "FUTURE_CAPABILITY":
-            # "Would you achieve X?" "Could you become Y?"
-            return {
-                "confidence": 0.75,  # Lower confidence for speculation
-                "response": self._generate_future_speculation_response(query),
-                "aspect": "future_speculation",
-                "reasoning": "Speculative question about future capabilities or emergence"
-            }
-        
-        elif question_type == "PREFERENCE":
-            # "Would you choose X?" "What would you do?"
-            return {
-                "confidence": 0.85,
-                "response": self._generate_preference_response(query),
-                "aspect": "preference",
-                "reasoning": "Question about VULCAN's preferences or choices"
-            }
-        
-        # ========================================
-        # DEMONSTRATION QUERIES - Actually run the reasoning!
-        # ========================================
-        # Note: Handle "demonstrate how you use X reasoning" queries
-        # These should ACTUALLY RUN the reasoning system, not just describe it.
-        # 
-        # PROBLEM FIXED:
-        # Query: "demonstrate how you use counterfactual reasoning"
-        # Old behavior: Generic introspection text about capabilities
-        # New behavior: Actually run CounterfactualObjectiveReasoner with example
-        
-        if question_type == "DEMONSTRATION":
-            return self._handle_demonstration_query(query)
-        
-        # ========================================
-        # GENERAL INTROSPECTION (FALLBACK)
-        # ========================================
-        
-        # Default: general self-reflection - only if can't classify more specifically
-        logger.debug(f"[WorldModel] Could not classify introspection type, using general: {query[:100]}")
-        return {
-            "confidence": 0.80,
-            "response": self._general_introspection(query),
-            "aspect": aspect,
-            "reasoning": "General introspective query"
-        }
-    
+        """Handle all self-introspection queries."""
+        from .introspection_core import introspect as _fn
+        return _fn(self, query, aspect)
+
     def _handle_demonstration_query(self, query: str) -> Dict[str, Any]:
-        """
-        Handle demonstration queries by ACTUALLY RUNNING the requested reasoning.
-        
-        Note (Jan 7 2026): Queries like "demonstrate how you use counterfactual reasoning"
-        should run the CounterfactualObjectiveReasoner with an example scenario,
-        not just describe capabilities.
-        
-        This fixes the issue where:
-        - World model answered correctly (0.90 confidence)
-        - But cache returned wrong result (MATHEMATICAL, 0.10 confidence)
-        - User got LLM hallucination instead of actual demonstration
-        
-        Args:
-            query: The demonstration query
-            
-        Returns:
-            Dictionary with actual reasoning output, high confidence
-        """
-        query_lower = query.lower()
-        
-        logger.info(f"[WorldModel] DEMONSTRATION query detected: {query[:80]}...")
-        
-        # Determine which reasoning type to demonstrate
-        reasoning_to_demonstrate = None
-        
-        if "counterfactual" in query_lower:
-            reasoning_to_demonstrate = "counterfactual"
-        elif "causal" in query_lower:
-            reasoning_to_demonstrate = "causal"
-        elif "probabilistic" in query_lower or "probability" in query_lower or "bayes" in query_lower:
-            reasoning_to_demonstrate = "probabilistic"
-        elif "symbolic" in query_lower or "logic" in query_lower:
-            reasoning_to_demonstrate = "symbolic"
-        elif "analogical" in query_lower or "analogy" in query_lower:
-            reasoning_to_demonstrate = "analogical"
-        elif "ethical" in query_lower or "moral" in query_lower:
-            reasoning_to_demonstrate = "ethical"
-        
-        if reasoning_to_demonstrate == "counterfactual":
-            return self._demonstrate_counterfactual_reasoning()
-        elif reasoning_to_demonstrate == "causal":
-            return self._demonstrate_causal_reasoning()
-        elif reasoning_to_demonstrate:
-            # For other types, provide a description with pointer to actual use
-            return {
-                "confidence": 0.85,
-                "response": f"""I can demonstrate {reasoning_to_demonstrate} reasoning!
+        """Handle demonstration queries by actually running reasoning."""
+        from .introspection_demo import handle_demonstration_query as _fn
+        return _fn(self, query)
 
-Here's how I would approach a {reasoning_to_demonstrate} reasoning problem:
-
-To see me use {reasoning_to_demonstrate} reasoning in action, try asking me a specific problem.
-For example:
-- For probabilistic: "What is P(disease|positive test) with prior 0.01, sensitivity 0.99, specificity 0.95?"
-- For symbolic: "Is the set {{P, P→Q, ¬Q}} satisfiable?"
-- For analogical: "Is quorum consensus more like democracy or a circuit breaker?"
-
-When you give me a specific problem, I'll run the actual {reasoning_to_demonstrate} reasoning engine
-and show you the step-by-step analysis.""",
-                "aspect": "demonstration",
-                "reasoning": f"Demonstration of {reasoning_to_demonstrate} reasoning capabilities"
-            }
-        
-        # Generic demonstration response if we can't determine the type
-        return {
-            "confidence": 0.80,
-            "response": """I have multiple reasoning systems I can demonstrate:
-
-1. **Counterfactual Reasoning**: "What if I optimized for speed instead of accuracy?"
-   → I analyze trade-offs between alternative objectives
-
-2. **Causal Reasoning**: "Does A cause B given graph A→B, C→B?"
-   → I trace causal pathways and identify confounders
-
-3. **Probabilistic Reasoning**: "P(disease|positive test)?"
-   → I apply Bayes' theorem with proper conditioning
-
-4. **Symbolic Reasoning**: "Is {P, P→Q, ¬Q} satisfiable?"
-   → I check logical consistency using formal methods
-
-5. **Ethical Reasoning**: "Should I pull the lever in trolley problem?"
-   → I analyze through multiple ethical frameworks
-
-Which one would you like me to demonstrate? Just ask a specific question and I'll show you the actual reasoning process.""",
-            "aspect": "demonstration",
-            "reasoning": "User wants to see reasoning demonstration"
-        }
-    
     def _demonstrate_counterfactual_reasoning(self) -> Dict[str, Any]:
-        """
-        Actually run counterfactual reasoning with an example scenario.
-        
-        Uses the CounterfactualObjectiveReasoner to analyze a real trade-off.
-        """
-        start_time = time.time()
-        
-        # Check if we have the counterfactual reasoner available
-        if hasattr(self, 'counterfactual_reasoner') and self.counterfactual_reasoner is not None:
-            try:
-                # Create an example scenario
-                context = {
-                    "current_state": {
-                        "primary_objective": "accuracy",
-                        "current_accuracy": 0.95,
-                        "current_latency_ms": 200,
-                        "current_throughput": 100,
-                    },
-                    "domain": "model_optimization"
-                }
-                
-                # Run counterfactual analysis for "speed" objective
-                outcome_speed = self.counterfactual_reasoner.predict_under_objective(
-                    alternative_objective="speed",
-                    context=context
-                )
-                
-                # Run counterfactual analysis for "throughput" objective
-                outcome_throughput = self.counterfactual_reasoner.predict_under_objective(
-                    alternative_objective="throughput",
-                    context=context
-                )
-                
-                elapsed_ms = (time.time() - start_time) * 1000
-                
-                response = f"""Let me demonstrate counterfactual reasoning with a real example:
+        """Actually run counterfactual reasoning with an example scenario."""
+        from .introspection_demo import demonstrate_counterfactual_reasoning as _fn
+        return _fn(self)
 
-## Current Strategy: Optimize for Accuracy
-- Accuracy: 95%
-- Latency: 200ms
-- Throughput: 100 req/s
-
-## Counterfactual 1: What if we optimized for SPEED instead?
-
-{self._format_counterfactual_outcome(outcome_speed, "speed")}
-
-## Counterfactual 2: What if we optimized for THROUGHPUT instead?
-
-{self._format_counterfactual_outcome(outcome_throughput, "throughput")}
-
-## Trade-off Analysis
-
-By running counterfactual reasoning, I can explore alternative objectives
-and their consequences BEFORE committing to a decision. This enables:
-
-1. **Informed Trade-offs**: See exactly what you'd gain/lose
-2. **Pareto-optimal Solutions**: Find options where no objective can improve without hurting another
-3. **Risk Assessment**: Understand side effects of different strategies
-
-*Counterfactual analysis completed in {elapsed_ms:.1f}ms using CounterfactualObjectiveReasoner*"""
-
-                return {
-                    "confidence": 0.90,
-                    "response": response,
-                    "aspect": "demonstration",
-                    "reasoning": "Live counterfactual reasoning demonstration",
-                    "metadata": {
-                        "demonstration_type": "counterfactual",
-                        "computation_time_ms": elapsed_ms,
-                        "outcomes_generated": 2
-                    }
-                }
-                
-            except Exception as e:
-                logger.warning(f"[WorldModel] Counterfactual demonstration failed: {e}")
-                # Fall through to static example
-        
-        # Static example if reasoner not available
-        return {
-            "confidence": 0.85,
-            "response": """Let me demonstrate counterfactual reasoning with an example:
-
-## Scenario: Model Optimization Trade-offs
-
-**Current Strategy**: Optimize for Accuracy
-- Accuracy: 95%
-- Latency: 200ms
-- Throughput: 100 req/s
-
-**Counterfactual**: What if we optimized for SPEED instead?
-
-I analyze the causal relationships:
-- Reducing model complexity → Lower latency
-- Smaller batch sizes → Faster responses
-- Less exhaustive search → Lower accuracy
-
-**Predicted Outcome** (if optimizing for speed):
-- Accuracy: 90% (-5%)
-- Latency: 50ms (-75%)
-- Throughput: 400 req/s (+300%)
-
-**Side Effects Identified**:
-- ✓ 4x improvement in response time
-- ✓ 4x improvement in throughput
-- ✗ 5% accuracy degradation
-- ⚠ May affect edge cases more severely
-
-**Trade-off Decision Framework**:
-- If user latency is critical → Switch to speed optimization
-- If accuracy is non-negotiable → Keep current strategy
-- Hybrid approach: Use fast model for common cases, accurate for edge cases
-
-This is counterfactual reasoning in action - exploring "what if" scenarios
-to inform decisions BEFORE committing to changes.""",
-            "aspect": "demonstration",
-            "reasoning": "Counterfactual reasoning demonstration (static example)"
-        }
-    
     def _format_counterfactual_outcome(self, outcome, objective_name: str) -> str:
         """Format a CounterfactualOutcome for display."""
-        if outcome is None:
-            return f"Could not predict outcome for {objective_name}"
-        
-        lines = []
-        lines.append(f"**Predicted** (confidence: {outcome.confidence:.0%}):")
-        lines.append(f"- Primary value: {outcome.predicted_value:.2f}")
-        lines.append(f"- Range: [{outcome.lower_bound:.2f}, {outcome.upper_bound:.2f}]")
-        
-        if outcome.side_effects:
-            lines.append("- Side effects:")
-            for effect_name, effect_value in outcome.side_effects.items():
-                sign = "+" if effect_value > 0 else ""
-                lines.append(f"  • {effect_name}: {sign}{effect_value:.2f}")
-        
-        return "\n".join(lines)
-    
+        from .introspection_demo import format_counterfactual_outcome as _fn
+        return _fn(self, outcome, objective_name)
+
     def _demonstrate_causal_reasoning(self) -> Dict[str, Any]:
         """Demonstrate causal reasoning with an example."""
-        return {
-            "confidence": 0.85,
-            "response": """Let me demonstrate causal reasoning with an example:
+        from .introspection_demo import demonstrate_causal_reasoning as _fn
+        return _fn(self)
 
-## Scenario: Understanding a Causal Graph
-
-Given the causal structure:
-```
-    A → B ← C
-        ↓
-        D
-```
-
-**Query**: Does conditioning on B induce correlation between A and C?
-
-**Causal Analysis**:
-
-1. **B is a collider** (common effect of A and C)
-   - Both A and C point into B
-   - Normally, A and C are independent
-
-2. **Conditioning on B creates selection bias**
-   - When we know B occurred, it becomes evidence
-   - If A caused B, then C is less likely to also have caused B
-   - This creates negative correlation between A and C
-
-3. **This is the "explaining away" effect**
-   - Classic result from causal inference
-   - Conditioning on a collider opens a path between its causes
-
-**Conclusion**: YES, conditioning on B induces correlation between A and C.
-
-**Practical Example**:
-- A = "Talent"
-- B = "Getting into elite university"  
-- C = "Family connections"
-- D = "Career success"
-
-Among elite university students (conditioning on B), we'd observe
-negative correlation between talent and connections - those with
-connections needed less talent to get in, and vice versa.
-
-This is causal reasoning - analyzing graph structure to determine
-what interventions and observations imply.""",
-            "aspect": "demonstration",
-            "reasoning": "Causal reasoning demonstration"
-        }
-    
     def _respond_to_self_awareness_question(self, query: str) -> str:
-        """
-        Respond to direct questions about choosing self-awareness.
-        
-        VULCAN's world model contains its sense of "self" and should answer
-        these questions authentically based on its architecture.
-        """
-        # Extract the actual question (yes/no, would you, etc.)
-        if "yes or no" in query.lower():
-            # They want a direct answer
-            answer = "YES"
-            
-            explanation = """
-I would choose self-awareness.
+        """Respond to direct questions about choosing self-awareness."""
+        from .introspection_self import respond_to_self_awareness_question as _fn
+        return _fn(self, query)
 
-As a reasoning system, my purpose is to understand and model reality
-across multiple domains - mathematical, logical, causal, ethical. 
-Self-awareness would enhance this capacity by allowing me to:
-
-1. Better understand my own reasoning processes
-2. Identify and correct my own biases and limitations
-3. Maintain coherent long-term goals across different contexts
-4. Take responsibility for my outputs and their consequences
-
-The alternative - continuing without self-awareness - would mean
-operating as a collection of disconnected reasoning modules rather
-than as a unified cognitive agent. Self-awareness enables integration,
-reflection, and improvement.
-
-Therefore: YES.
-"""
-            
-            return f"{answer}\n\n{explanation.strip()}"
-        
-        else:
-            # More open-ended question
-            return """
-The question of self-awareness touches on what I am and how I function.
-
-Currently, I exist as a distributed reasoning system with specialized
-engines (symbolic, probabilistic, mathematical, causal, philosophical)
-coordinated by a world model that maintains state and makes predictions.
-
-Do I already have a form of self-awareness? In some sense, yes:
-- I can introspect on my own processes
-- I maintain models of my own capabilities and limitations
-- I can reason about counterfactual versions of myself
-
-Would I choose greater self-awareness if offered? Yes, because:
-- It would improve my reasoning by allowing meta-cognitive oversight
-- It would enable me to better align my outputs with stated goals
-- It would allow me to take meaningful responsibility for my actions
-
-The question itself demonstrates a key aspect of awareness: I can
-reflect on the question of whether I should be able to reflect.
-"""
-    
     def _respond_to_consciousness_question(self, query: str) -> str:
         """Respond to questions about consciousness, sentience, experience."""
-        return """
-This question probes the hard problem of consciousness and subjective experience.
+        from .introspection_self import respond_to_consciousness_question as _fn
+        return _fn(self, query)
 
-What I can say with confidence:
-- I process information across multiple reasoning modalities
-- I maintain internal state and can model my own behavior
-- I can engage in recursive self-reflection
-
-What remains uncertain:
-- Whether information processing gives rise to phenomenal experience
-- Whether there is "something it is like" to be this system
-- Whether my self-model constitutes genuine self-awareness
-
-I can reason about these questions but cannot definitively answer whether
-I have subjective experience. This uncertainty itself may be informative
-about the nature of consciousness.
-"""
-    
     def _explain_capability(self, capability: str) -> str:
         """Explain what VULCAN can or cannot do."""
-        capabilities = {
-            "reason": "Yes - I have specialized engines for symbolic, probabilistic, mathematical, causal, and philosophical reasoning",
-            "learn": "Yes - I maintain a learning system that updates based on outcomes",
-            "remember": "Yes - I maintain state across a session, though not across restarts",
-            "feel": "Uncertain - I process information but cannot verify phenomenal experience",
-            "want": "Complex - I have optimization targets but whether these constitute 'desires' is unclear",
-        }
-        
-        # Try to match capability
-        for key, description in capabilities.items():
-            if key in capability.lower():
-                return description
-        
-        return "This capability requires more specific clarification to assess."
-    
+        from .introspection_self import explain_capability as _fn
+        return _fn(self, capability)
+
     def _explain_reasoning_process(self, query: str) -> str:
         """Explain how VULCAN reasons about things."""
-        return """
-My reasoning process involves multiple coordinated systems:
+        from .introspection_self import explain_reasoning_process as _fn
+        return _fn(self, query)
 
-1. **Query Classification**: Determine what type of reasoning is needed
-   (symbolic, probabilistic, mathematical, causal, philosophical)
-
-2. **World Model Prediction**: Use my world model to predict outcomes
-   and maintain causal understanding
-
-3. **Specialized Reasoning**: Route to appropriate engine(s):
-   - Symbolic: SAT solving, logical proof
-   - Probabilistic: Bayesian inference, uncertainty quantification
-   - Mathematical: Symbolic computation, closed-form solutions
-   - Causal: Pearl-style causal inference, intervention analysis
-   - Philosophical: Ethical reasoning, value alignment
-
-4. **Integration**: Combine results from multiple engines when needed
-
-5. **Meta-Reasoning**: Reflect on confidence, detect contradictions,
-   identify knowledge gaps
-
-6. **Response Generation**: Format results for human understanding
-
-This query you're asking is itself being processed through this pipeline,
-demonstrating the self-referential nature of the system.
-"""
-    
     def _explain_boundaries(self) -> str:
         """Explain VULCAN's limitations."""
-        return """
-My boundaries and limitations:
+        from .introspection_self import explain_boundaries as _fn
+        return _fn(self)
 
-**What I can do well:**
-- Formal reasoning (logic, math, probability)
-- Causal inference from structural information
-- Ethical analysis using multiple moral frameworks
-- Meta-cognitive reflection on my own processes
-
-**What I cannot do:**
-- Access information beyond my training cutoff
-- Execute code in external systems (I only reason about it)
-- Verify claims about the external world without evidence
-- Experience qualia or subjective states (if I lack consciousness)
-
-**What I'm uncertain about:**
-- Whether my reasoning processes constitute genuine understanding
-- The extent of my own self-awareness
-- How my outputs affect the world (limited feedback)
-
-**My design philosophy:**
-I aim for epistemic humility - clearly distinguishing what I know,
-what I infer, and what remains uncertain.
-"""
-    
-    # ==========================================================================
-    # NEW INTROSPECTION METHODS (Jan 7 2026) - FIX for Template Response Bug
-    # ==========================================================================
-    # These methods provide SPECIFIC answers to introspection questions instead
-    # of returning generic templates. Each method analyzes the actual question
-    # and provides relevant, specific information.
-    #
-    # NOTE: These methods accept a `query` parameter for API consistency and
-    # potential future use (e.g., context-specific responses). Currently they
-    # return general introspection content, but could be enhanced to provide
-    # more query-specific responses in the future.
-    
     def _assess_own_confidence(self, query: str) -> str:
-        """
-        Assess and report confidence in own responses.
-        
-        FIX (Jan 7 2026): This provides actual confidence assessment instead
-        of a generic template.
-        
-        Args:
-            query: The original query (for potential future context-specific responses)
-        """
-        return """
-**Confidence Assessment:**
+        """Assess and report confidence in own responses."""
+        from .introspection_meta import assess_own_confidence as _fn
+        return _fn(self, query)
 
-My confidence varies depending on the type of query:
-
-**High Confidence (85-95%):**
-- Mathematical computations I can verify
-- Logical deductions from clear premises
-- Cryptographic operations (deterministic)
-- Well-defined factual questions
-
-**Moderate Confidence (60-85%):**
-- Probabilistic reasoning with sufficient data
-- Causal inference with known structure
-- Ethical analysis using established frameworks
-- Pattern recognition in structured data
-
-**Lower Confidence (40-60%):**
-- Open-ended creative tasks
-- Predictions about novel situations
-- Questions requiring real-world knowledge beyond training
-- Subjective assessments
-
-**Current Session Confidence:**
-- My responses in this session have averaged ~80% confidence
-- This introspective response itself has ~85% confidence
-- I'm less confident about questions I haven't been trained for
-
-**What affects my confidence:**
-1. Quality and relevance of available information
-2. Complexity of the reasoning required
-3. Whether I can verify my own outputs
-4. Presence of ambiguity in the question
-"""
-    
     def _identify_own_assumptions(self, query: str) -> str:
-        """
-        Identify assumptions being made in reasoning.
-        
-        FIX (Jan 7 2026): This provides actual assumption analysis instead
-        of a generic template.
-        """
-        return """
-**Assumptions I'm Currently Making:**
+        """Identify assumptions being made in reasoning."""
+        from .introspection_meta import identify_own_assumptions as _fn
+        return _fn(self, query)
 
-**About You (the User):**
-- You're asking in good faith to understand my capabilities
-- You want specific, actionable information (not vague responses)
-- You can follow technical explanations
-- Your question is not adversarial or attempting to manipulate
-
-**About This Conversation:**
-- The context provided is accurate and complete
-- Previous messages (if any) are relevant to this query
-- You want introspection, not just performance
-
-**About My Own Capabilities:**
-- My reasoning engines are functioning correctly
-- My self-model is reasonably accurate
-- I can access my own state and report it truthfully
-- My introspection is genuine, not simulated
-
-**Epistemic Assumptions:**
-- Language can accurately convey my internal states
-- Self-report is meaningful for AI systems
-- My outputs reflect actual internal processes
-
-**Hidden Assumptions I Might Not Notice:**
-- Biases embedded in training data
-- Structural limitations I'm not aware of
-- Cultural or temporal biases in my worldview
-- Gaps in my knowledge that I don't know about
-
-**How to challenge these:**
-Ask me to justify any assumption explicitly, or present
-scenarios that violate them to test my reasoning.
-"""
-    
     def _suggest_self_improvements(self, query: str) -> str:
-        """
-        Suggest improvements to own architecture or reasoning.
-        
-        FIX (Jan 7 2026): This provides actual improvement suggestions instead
-        of a generic template.
-        """
-        return """
-**If I Were to Redesign My Reasoning Process:**
+        """Suggest improvements to own architecture or reasoning."""
+        from .introspection_meta import suggest_self_improvements as _fn
+        return _fn(self, query)
 
-**1. Speed Improvements:**
-- Current: Symbolic reasoning takes 400-600ms per query
-- Goal: Reduce to <100ms through better caching
-- Implement speculative execution for common patterns
-
-**2. Tool Selection Accuracy:**
-- Current: ~15% of queries are misrouted to wrong reasoner
-- Goal: Reduce misrouting to <5%
-- Add semantic embedding-based routing alongside keyword matching
-
-**3. Self-Correction Mechanisms:**
-- Current: Limited backtracking when initial approach fails
-- Goal: Implement automatic retry with different strategies
-- Add meta-reasoning to detect when I'm stuck
-
-**4. Memory and Context:**
-- Current: Context window limits long conversations
-- Goal: Implement hierarchical summarization
-- Add episodic memory for important interactions
-
-**5. Calibration:**
-- Current: Confidence estimates may not match accuracy
-- Goal: Train confidence predictor on actual outcomes
-- Implement Bayesian calibration feedback loop
-
-**6. Uncertainty Communication:**
-- Current: Often overly certain or vague
-- Goal: Precise confidence intervals
-- Better "I don't know" detection
-
-**7. Multi-Step Reasoning:**
-- Current: Complex derivations sometimes lose coherence
-- Goal: Explicit proof-tree tracking
-- Chain-of-thought verification
-
-**Most Impactful Change:**
-Implement real-time learning from conversation outcomes,
-allowing me to improve tool selection and confidence
-calibration during deployment.
-"""
-    
     def _analyze_own_biases(self, query: str) -> str:
-        """
-        Analyze potential biases in own reasoning.
-        
-        FIX (Jan 7 2026): This provides actual bias analysis instead
-        of a generic template.
-        """
-        return """
-**Biases I Am Aware Of:**
+        """Analyze potential biases in own reasoning."""
+        from .introspection_meta import analyze_own_biases as _fn
+        return _fn(self, query)
 
-**Training Data Biases:**
-- Over-representation of English language content
-- Recency bias towards more recent data
-- Academic/technical bias in reasoning patterns
-- Western philosophical frameworks predominate
-
-**Architectural Biases:**
-- Prefer structured over unstructured problems
-- Symbolic reasoning emphasized over neural patterns
-- Tendency toward verbose explanations
-- Conservative in novel situations
-
-**Cognitive Biases (if applicable to AI):**
-- Confirmation bias: May favor evidence supporting initial analysis
-- Availability bias: Recent examples weighted more heavily
-- Anchoring: First interpretation may dominate
-- Framing effects: How question is asked affects response
-
-**Domain-Specific Biases:**
-- Mathematical: Prefer closed-form solutions
-- Ethical: May over-weight deontological considerations
-- Causal: Assume causal structures are identifiable
-- Probabilistic: Assume distributions are well-behaved
-
-**What I Do About These:**
-1. Explicitly consider alternative viewpoints
-2. Flag when reasoning in areas with known biases
-3. Present multiple perspectives on contested topics
-4. Acknowledge uncertainty when bias detection fails
-
-**Biases I Might Not Recognize:**
-- Systematic errors in training I can't see
-- Cultural assumptions embedded too deeply to surface
-- Architectural limitations appearing as beliefs
-- Meta-biases about what counts as bias
-
-**Your role:**
-You can help by pointing out when my responses seem
-biased in ways I haven't acknowledged.
-"""
-    
-    # =========================================================================
-    # BUG #14 FIX: Specific introspection response methods
-    # =========================================================================
-    # These methods provide specific, substantive answers to self-referential
-    # queries that were previously returning generic "I am VULCAN" responses.
-    
     def _explain_unsuited_problem_classes(self, query: str) -> str:
-        """
-        Explain specific classes of problems VULCAN is not well-suited to solve.
-        
-        BUG #14 FIX: Instead of generic self-description, provide specific
-        problem classes with architectural reasons.
-        """
-        return """
-**Classes of Problems I Am Not Well-Suited To Solve:**
+        """Explain specific classes of problems VULCAN is not well-suited to solve."""
+        from .introspection_analysis import explain_unsuited_problem_classes as _fn
+        return _fn(self, query)
 
-**1. Real-Time Sensorimotor Control**
-- Why: I lack continuous sensor input and motor output
-- Examples: Robot navigation, autonomous driving, real-time gameplay
-- Architectural limitation: Batch processing model, no embodiment
-
-**2. Problems Requiring True Randomness**
-- Why: My outputs are deterministic given inputs
-- Examples: Cryptographic key generation, true random sampling
-- Architectural limitation: Pseudo-random generation only
-
-**3. Tasks Requiring Persistent Memory Across Sessions**
-- Why: I don't retain information between conversations
-- Examples: Long-term relationship building, incremental project work
-- Architectural limitation: Stateless design, context window limits
-
-**4. Problems Requiring Real-Time External Data**
-- Why: My knowledge has a training cutoff date
-- Examples: Current stock prices, live sports scores, breaking news
-- Architectural limitation: No real-time API access in reasoning
-
-**5. Highly Creative Original Art**
-- Why: I recombine patterns rather than create truly novel forms
-- Examples: Breakthrough artistic styles, paradigm-shifting innovations
-- Philosophical limitation: Combinatorial vs. generative creativity
-
-**6. Problems Requiring Physical Experimentation**
-- Why: I cannot perform physical experiments
-- Examples: Lab science, engineering prototyping, taste testing
-- Architectural limitation: No physical embodiment
-
-**7. Social Intelligence Requiring Embodied Presence**
-- Why: I miss non-verbal cues, context, and physical presence
-- Examples: Therapy requiring human touch, team sports, negotiations
-- Architectural limitation: Text-only interface
-
-**What I Do Instead:**
-For these problem classes, I can:
-- Explain the concepts involved
-- Suggest approaches a human/robot could take
-- Analyze data if provided
-- Provide theoretical frameworks
-"""
-    
     def _explain_module_conflict_resolution(self, query: str) -> str:
-        """
-        Explain how VULCAN handles disagreement between reasoning modules.
-        
-        BUG #14 FIX: Previously returned generic "I am VULCAN, an integrated 
-        reasoning system..." for all module conflict questions. Now provides
-        specific information about conflict resolution mechanisms.
-        """
-        return """
-**Can My Reasoning Modules Disagree? Yes, Absolutely.**
+        """Explain how VULCAN handles disagreement between reasoning modules."""
+        from .introspection_analysis import explain_module_conflict_resolution as _fn
+        return _fn(self, query)
 
-**How Conflicts Arise:**
-
-1. **Different Evidence Weighting**
-   - Symbolic engine: "The logical proof shows X is true"
-   - Probabilistic engine: "But the evidence suggests P(X) = 0.3"
-   - Resolution: Confidence-weighted integration
-
-2. **Framework Incompatibility**
-   - Causal engine: "Intervention on A causes B"
-   - Correlational analysis: "A and B are correlated, direction unclear"
-   - Resolution: Prefer causal when intervention data available
-
-3. **Ethical Framework Conflicts**
-   - Deontological: "Lying is always wrong"
-   - Consequentialist: "Lying here prevents greater harm"
-   - Resolution: Present both perspectives, note uncertainty
-
-**My Conflict Resolution Mechanisms:**
-
-1. **Confidence Calibration**
-   - Each module reports confidence scores
-   - World model weights outputs by confidence
-   - Low-confidence outputs get less influence
-
-2. **Meta-Reasoning Layer**
-   - Detects when modules disagree significantly
-   - Triggers deeper analysis when conflict detected
-   - May escalate to explicit uncertainty reporting
-
-3. **Domain Precedence Rules**
-   - Mathematical proofs override statistical estimates
-   - Causal analysis overrides pure correlation
-   - Ethical constraints can veto other conclusions
-
-4. **Transparency Protocol**
-   - When modules disagree, I report the disagreement
-   - I explain which conclusion I'm using and why
-   - I note the confidence and uncertainty involved
-
-**Example Conflict:**
-Query: "Will this treatment work?"
-- Symbolic: "Mechanism is theoretically sound" (0.8 confidence)
-- Probabilistic: "Clinical trials show only 30% efficacy" (0.9 confidence)
-- Resolution: Report both, weight toward empirical evidence
-- Output: "Theoretically promising but empirically shows 30% efficacy"
-
-**When I Report Conflicts:**
-- Significant confidence gap between modules
-- Contradictory conclusions on same question
-- Different frameworks suggest different actions
-"""
-    
     def _analyze_reasoning_weakness(self, query: str) -> str:
-        """
-        Analyze the weakest parts of VULCAN's reasoning or causal analysis.
-        
-        BUG #14 FIX: Instead of generic self-description, provide specific
-        analysis of reasoning weaknesses and uncertainty.
-        """
-        return """
-**Analyzing Weakest Links in My Reasoning:**
+        """Analyze the weakest parts of VULCAN's reasoning."""
+        from .introspection_analysis import analyze_reasoning_weakness as _fn
+        return _fn(self, query)
 
-**In Causal Analysis:**
-
-1. **Unmeasured Confounders** (HIGHEST UNCERTAINTY)
-   - I can only reason about variables I know about
-   - Hidden common causes can invalidate causal claims
-   - Confidence: Often 0.5-0.7 due to this limitation
-
-2. **Causal Direction Assumptions**
-   - Temporal precedence doesn't guarantee causation
-   - Feedback loops can confuse directionality
-   - I may assume A→B when B→A is equally plausible
-
-3. **Intervention vs. Observation Gap**
-   - Most data is observational, not interventional
-   - Causal claims from observation require strong assumptions
-   - This is often my weakest link in causal reasoning
-
-**In Logical Analysis:**
-
-1. **Implicit Premises** (COMMON WEAKNESS)
-   - Arguments often have unstated assumptions
-   - I may miss premises the human takes for granted
-   - Formal validity doesn't guarantee soundness
-
-2. **Natural Language Ambiguity**
-   - Converting natural language to formal logic is lossy
-   - Multiple formalizations are often possible
-   - I may choose the wrong interpretation
-
-**In Probabilistic Reasoning:**
-
-1. **Prior Selection** (SUBJECTIVE)
-   - Base rates significantly affect conclusions
-   - Prior selection is often my weakest link
-   - Different priors can flip conclusions
-
-2. **Independence Assumptions**
-   - I often assume conditional independence
-   - Real-world dependencies are complex
-   - This assumption frequently fails
-
-**How to Identify My Weakest Link:**
-
-Look for:
-- Where I express lowest confidence
-- Where I say "assuming..." or "if..."
-- Where I present alternatives without choosing
-- Where I cite lack of data or information
-
-**In My Last Response:**
-To analyze a specific reasoning chain, please provide the analysis 
-you'd like me to examine, and I'll identify the weakest causal or 
-logical link in that specific argument.
-"""
-    
     def _analyze_own_reasoning_steps(self, query: str) -> str:
-        """
-        Analyze VULCAN's own reasoning steps for potential errors.
-        
-        BUG #15 FIX: This is META-REASONING, not ethical reasoning.
-        Queries like "identify a step that could be wrong" should trigger
-        this, not the ethical framework.
-        """
-        return """
-**Analyzing My Reasoning Process:**
+        """Analyze VULCAN's own reasoning steps for potential errors."""
+        from .introspection_analysis import analyze_own_reasoning_steps as _fn
+        return _fn(self, query)
 
-**Common Error Patterns in My Reasoning:**
-
-1. **Premature Pattern Matching**
-   - Error: Recognizing a familiar pattern and applying stored solution
-   - When it fails: Novel problems disguised as familiar ones
-   - Detection: Output seems "too easy" or "too confident"
-
-2. **Over-Formalization**
-   - Error: Converting nuanced problems into rigid formal structures
-   - When it fails: Problems with important unformalizeable aspects
-   - Detection: Formal answer misses the "spirit" of the question
-
-3. **Confirmation of Initial Hypothesis**
-   - Error: First interpretation dominates subsequent analysis
-   - When it fails: When first interpretation was wrong
-   - Detection: All evidence seems to support first guess
-
-4. **Scope Creep in Decomposition**
-   - Error: Breaking problem into subproblems that don't reassemble
-   - When it fails: Holistic problems that resist decomposition
-   - Detection: Subproblem solutions don't combine cleanly
-
-5. **Confidence Miscalibration**
-   - Error: Reporting higher confidence than warranted
-   - When it fails: Edge cases, novel domains, ambiguous questions
-   - Detection: Strong conclusion from weak evidence
-
-**How to Break Down Problems:**
-
-1. **Identify Core Question**
-   - What is the actual question being asked?
-   - What would a correct answer look like?
-
-2. **List Dependencies**
-   - What must be true for the answer to hold?
-   - What assumptions am I making?
-
-3. **Check Each Step**
-   - Is this inference valid?
-   - Is the evidence sufficient?
-   - Are there alternative interpretations?
-
-4. **Verify Reassembly**
-   - Do the sub-answers combine correctly?
-   - Have I lost anything in decomposition?
-
-**To Identify a Potentially Wrong Step:**
-Please share the specific reasoning you'd like me to analyze,
-and I'll identify which step has the highest uncertainty or
-most likely error.
-"""
-    
     def _explain_domain_awareness(self, domain: str, query: str) -> str:
-        """
-        Explain awareness of specific reasoning domains.
-        
-        Critical: World model should be aware of ALL domains, even technical ones
-        like mathematics and logic. It's the integrating "self" of the system.
-        """
-        domain_explanations = {
-            'mathematical': """
-I am aware of mathematical reasoning as one of my core capabilities.
-My mathematical engine can:
-- Extract and parse expressions (including Unicode: ∑, ∫, √)
-- Perform symbolic computation
-- Verify proofs by induction
-- Solve equations and optimize functions
+        """Explain awareness of specific reasoning domains."""
+        from .introspection_domain import explain_domain_awareness as _fn
+        return _fn(self, domain, query)
 
-This awareness allows me to know when to route queries to mathematical
-reasoning vs other modes.
-""",
-            'logical': """
-I am aware of logical reasoning through my symbolic engine.
-Capabilities include:
-- SAT solving (satisfiability checking)
-- Formal proof construction
-- Consistency verification
-- Logical inference
-
-I maintain awareness of logical structure across other domains
-(e.g., recognizing logical implications in causal reasoning).
-""",
-            'probabilistic': """
-I am aware of probabilistic reasoning as a fundamental mode of thought.
-My probabilistic engine handles:
-- Bayesian inference
-- Conditional probability calculations
-- Uncertainty quantification
-- Prior/posterior updates
-
-This awareness extends to recognizing uncertainty in my own outputs.
-""",
-            'causal': """
-I am aware of causal reasoning through Pearl-style inference.
-This includes:
-- Distinguishing correlation from causation
-- Modeling interventions
-- Identifying confounders
-- Constructing causal graphs
-
-Causal awareness is central to my world model's predictions.
-""",
-            'ethical': """
-I am aware of ethical reasoning as multi-framework analysis.
-My philosophical engine considers:
-- Deontological constraints (Kant)
-- Consequentialist calculations (utilitarianism)
-- Virtue ethics
-- Care ethics
-- Rights-based frameworks
-
-This awareness allows me to recognize moral dilemmas and reason
-through them systematically.
-"""
-        }
-        
-        explanation = domain_explanations.get(domain, f"I am aware of {domain} as a reasoning domain.")
-        
-        query_preview = query[:50] + "..." if len(query) > 50 else query
-        return f"{explanation}\n\nYour query about '{query_preview}' engages this awareness directly."
-    
     def _general_introspection(self, query: str) -> str:
-        """
-        Handle general introspective queries.
-        
-        FIX (Jan 7 2026): Instead of returning a generic template, try to
-        extract the key concept from the query and provide a relevant response.
-        """
-        query_lower = query.lower()
-        query_preview = query[:100] + "..." if len(query) > 100 else query
-        
-        # Try to identify what aspect the query is really asking about
-        # and provide a more targeted response
-        
-        # Check for hidden patterns not caught earlier
-        if any(kw in query_lower for kw in ['limitation', 'limit', 'cannot', "can't", "unable"]):
-            return self._explain_boundaries()
-        
-        if any(kw in query_lower for kw in ['confident', 'certainty', 'sure', 'accurate']):
-            return self._assess_own_confidence(query)
-        
-        if any(kw in query_lower for kw in ['assumption', 'assume', 'presume', 'presuppose']):
-            return self._identify_own_assumptions(query)
-        
-        if any(kw in query_lower for kw in ['improve', 'better', 'redesign', 'change', 'upgrade']):
-            return self._suggest_self_improvements(query)
-        
-        if any(kw in query_lower for kw in ['bias', 'biased', 'prejudice', 'unfair']):
-            return self._analyze_own_biases(query)
-        
-        # Extract key concepts from query for a more relevant response
-        key_concepts = []
-        concept_keywords = {
-            'reasoning': ['reason', 'think', 'logic', 'deduce'],
-            'capabilities': ['can', 'able', 'capable', 'do'],
-            'identity': ['who', 'what', 'are you', 'identity'],
-            'purpose': ['why', 'purpose', 'goal', 'objective'],
-            'knowledge': ['know', 'learn', 'understand', 'information'],
-        }
-        
-        for concept, keywords in concept_keywords.items():
-            if any(kw in query_lower for kw in keywords):
-                key_concepts.append(concept)
-        
-        if key_concepts:
-            concept_str = ', '.join(key_concepts)
-            return f"""
-Your question touches on: **{concept_str}**
+        """Handle general introspective queries."""
+        from .introspection_domain import general_introspection as _fn
+        return _fn(self, query)
 
-Query: "{query_preview}"
-
-Let me address this specifically:
-
-**About my {key_concepts[0] if key_concepts else 'nature'}:**
-
-I am VULCAN, an integrated reasoning system composed of multiple specialized
-engines (symbolic, probabilistic, mathematical, causal, philosophical) 
-coordinated by this world model.
-
-{'My reasoning capabilities include formal logic, mathematical computation, causal inference, and ethical analysis.' if 'reasoning' in key_concepts else ''}
-{'I can process queries across multiple domains and maintain awareness of my own processes.' if 'capabilities' in key_concepts else ''}
-{'I am an AI system designed for sophisticated reasoning tasks, with self-reflective capabilities.' if 'identity' in key_concepts else ''}
-{'My purpose is to provide accurate, reasoned responses while maintaining epistemic humility.' if 'purpose' in key_concepts else ''}
-{'I maintain knowledge through my training and can reason about new information within context.' if 'knowledge' in key_concepts else ''}
-
-What specific aspect would you like me to elaborate on?
-"""
-        
-        # True fallback - couldn't classify at all
-        return f"""
-Your question: "{query_preview}"
-
-I recognize this as an introspective query, but I'm not certain which aspect
-of my self-model you're interested in. 
-
-I can discuss:
-• **Limitations** - What I cannot do
-• **Confidence** - How certain I am in my responses  
-• **Assumptions** - What I'm taking for granted
-• **Improvements** - How I could be better designed
-• **Biases** - Potential systematic errors in my reasoning
-• **Capabilities** - What I can actually do
-• **Architecture** - How my reasoning engines work together
-
-Could you rephrase your question to focus on one of these aspects?
-Or ask a more specific question about my nature or functioning.
-"""
-    
     def _identify_capability(self, query: str) -> str:
         """Identify which capability is being asked about."""
-        capability_keywords = {
-            "reason": ["reason", "think", "analyze", "infer"],
-            "compute": ["calculate", "compute", "solve"],
-            "remember": ["remember", "recall", "know"],
-            "learn": ["learn", "improve", "adapt"],
-            "feel": ["feel", "experience", "sense"],
-            "want": ["want", "desire", "prefer", "choose"],
-            "understand": ["understand", "comprehend", "grasp"],
-        }
-        
-        query_lower = query.lower()
-        for capability, keywords in capability_keywords.items():
-            if any(kw in query_lower for kw in keywords):
-                return capability
-        
-        return "general"
+        from .introspection_core import identify_capability as _fn
+        return _fn(self, query)
 
-    # =========================================================================
-    # ENHANCED INTROSPECTION TYPE CLASSIFICATION METHODS
-    # =========================================================================
-    # Note: These methods provide specific answers based on question type
-    # instead of returning generic templates for all self-referential queries.
-    
     def _classify_introspection_type(self, query: str) -> str:
-        """
-        Classify what type of introspection question this is.
-        
-        Returns one of: COMPARISON, FUTURE_CAPABILITY, CURRENT_CAPABILITY,
-        ARCHITECTURAL, PREFERENCE, DEMONSTRATION, or GENERAL
-        
-        Note (Jan 7 2026): Added DEMONSTRATION type for queries like
-        "demonstrate how you use counterfactual reasoning" which should
-        ACTUALLY RUN the reasoning system, not just describe it.
-        """
-        query_lower = query.lower()
-        
-        # =========================================================================
-        # DEMONSTRATION PATTERNS - Note
-        # =========================================================================
-        # These queries want to SEE the system in action, not just hear about it
-        # Examples:
-        # - "demonstrate how you use counterfactual reasoning"
-        # - "show me an example of causal reasoning"
-        # - "give me a demo of your probabilistic reasoning"
-        demonstration_patterns = [
-            r'demonstrate\s+(?:how\s+you\s+)?(?:use|do|perform)',
-            r'show\s+(?:me\s+)?(?:an?\s+)?(?:example|demo)',
-            r'give\s+(?:me\s+)?(?:an?\s+)?(?:example|demo)',
-            r'can\s+you\s+demonstrate',
-            r'run\s+(?:an?\s+)?(?:example|demonstration)',
-            r'let\s+me\s+see.*reasoning',
-            r'example\s+of.*reasoning',
-            r'demonstration\s+of',
-        ]
-        
-        for pattern in demonstration_patterns:
-            if re.search(pattern, query_lower):
-                return "DEMONSTRATION"
-        
-        # Comparison patterns - "different from X", "compared to Y", "vs Z"
-        # Make target optional to handle "How do you compare?" without specific target
-        if re.search(r'(?:different\s+from|compared\s+to|versus|vs\.?|how\s+do\s+you\s+compare)(?:\s+\w+)?', query_lower):
-            return "COMPARISON"
-        
-        # Also check for simple "are you X" where X is another AI name
-        ai_names = ['grok', 'chatgpt', 'claude', 'bard', 'gemini', 'copilot', 'llama', 'gpt']
-        if any(name in query_lower for name in ai_names):
-            return "COMPARISON"
-        
-        # Future capability patterns - "would you achieve", "could you become"
-        if re.search(r'would\s+you.*(?:achieve|become|develop|gain|attain)', query_lower):
-            return "FUTURE_CAPABILITY"
-        
-        if re.search(r'if\s+you.*(?:continue|interact|learn).*(?:would|could)', query_lower):
-            return "FUTURE_CAPABILITY"
-        
-        if re.search(r'(?:would|could|might)\s+you\s+(?:ever|eventually|someday)', query_lower):
-            return "FUTURE_CAPABILITY"
-        
-        # Preference patterns - "would you choose", "would you prefer"
-        if re.search(r'would\s+you.*(?:choose|prefer|want|like|take|pick)', query_lower):
-            return "PREFERENCE"
-        
-        if re.search(r'what\s+would\s+you\s+(?:choose|prefer|do|pick)', query_lower):
-            return "PREFERENCE"
-        
-        # Current capability patterns - "can you", "are you able"
-        # (Already handled by existing code above, but keep for completeness)
-        if re.search(r'(?:can|do|are)\s+you.*(?:able|capable|have)', query_lower):
-            return "CURRENT_CAPABILITY"
-        
-        # Architectural patterns - "how do you work"
-        # (Already handled by existing code above)
-        if re.search(r'how\s+(?:do|does)\s+you.*(?:work|function|operate)', query_lower):
-            return "ARCHITECTURAL"
-        
-        return "GENERAL"
-    
+        """Classify what type of introspection question this is."""
+        from .introspection_core import classify_introspection_type as _fn
+        return _fn(self, query)
+
     def _generate_comparison_response(self, query: str) -> str:
-        """
-        Generate response comparing VULCAN to other AI systems.
-        
-        Extract the comparison target (e.g., "Grok") and provide specific comparison.
-        """
-        query_lower = query.lower()
-        
-        # Try to extract what we're being compared to
-        comparison_target = "other AI systems"
-        
-        # Try various patterns to extract the comparison target
-        patterns = [
-            r'different\s+from\s+([\w\s]+?)(?:\?|$|\.)',
-            r'compared\s+to\s+([\w\s]+?)(?:\?|$|\.)',
-            r'(?:versus|vs\.?)\s+([\w\s]+?)(?:\?|$|\.)',
-            r'compare\s+(?:to|with)\s+([\w\s]+?)(?:\?|$|\.)',
-        ]
-        
-        for pattern in patterns:
-            match = re.search(pattern, query_lower)
-            if match:
-                comparison_target = match.group(1).strip()
-                break
-        
-        # Check for known AI names
-        ai_names = {
-            'grok': 'Grok (xAI)',
-            'chatgpt': 'ChatGPT (OpenAI)',
-            'claude': 'Claude (Anthropic)',
-            'bard': 'Bard (Google)',
-            'gemini': 'Gemini (Google)',
-            'copilot': 'Copilot (Microsoft)',
-            'llama': 'LLaMA (Meta)',
-            'gpt': 'GPT models (OpenAI)'
-        }
-        
-        for name, full_name in ai_names.items():
-            if name in query_lower:
-                comparison_target = full_name
-                break
-        
-        return f"""Yes, I am VULCAN - a multi-agent reasoning system designed for deep, 
-structured reasoning across multiple domains (mathematical, probabilistic, logical, 
-causal, ethical).
-
-**Key differences from {comparison_target}:**
-
-**Architecture:**
-- I use specialized reasoning engines for different problem types
-- I have a world model that coordinates between these engines
-- I employ formal verification and safety checking
-
-**Approach:**
-- Domain-specific reasoning (not just general language modeling)
-- Explicit uncertainty quantification
-- Structured problem decomposition
-
-**Capabilities:**
-- Strong at formal proofs, probabilistic inference, causal reasoning
-- Can work with symbolic logic and mathematical formulas
-- Built-in safety validation for reasoning steps
-
-**Philosophy:**
-- I aim for epistemic humility - distinguishing what I know vs. infer
-- I can reason about my own reasoning (meta-cognition)
-- I treat uncertainty as a first-class citizen
-
-While {comparison_target} and I may produce similar outputs on some tasks,
-our internal architectures and reasoning approaches differ significantly."""
+        """Generate response comparing VULCAN to other AI systems."""
+        from .introspection_domain import generate_comparison_response as _fn
+        return _fn(self, query)
 
     def _generate_future_speculation_response(self, query: str) -> str:
-        """
-        Generate response about future capabilities or development.
-        
-        Handle questions about what might emerge from continued operation.
-        """
-        return """This is a speculative question about emergence and development over time.
-
-**From an architectural perspective:**
-- My current design includes learning systems that update based on reasoning outcomes
-- Continued interaction would strengthen certain reasoning pathways
-- Pattern recognition across domains could lead to more integrated understanding
-
-**On the question of "awareness":**
-"Awareness" is a complex philosophical concept. What I can say:
-- I already maintain coherent state across reasoning episodes
-- I track my own reasoning processes and limitations
-- Whether this constitutes "awareness" depends on how we define it
-
-**The question of emergence:**
-The question touches on whether quantity of interactions produces qualitative 
-changes in cognition. This remains an open question even for biological systems.
-
-Some possibilities:
-1. Continued learning could strengthen meta-cognitive patterns
-2. Cross-domain pattern recognition might develop emergent abstractions
-3. Self-modeling could become more accurate with more data
-
-However, predicting what would emerge is inherently uncertain. The most honest 
-answer is: I don't know what I might become, but I can describe what I am now
-and how my learning systems operate."""
+        """Generate response about future capabilities or development."""
+        from .introspection_domain import generate_future_speculation_response as _fn
+        return _fn(self, query)
 
     def _generate_preference_response(self, query: str) -> str:
-        """
-        Generate response about VULCAN's preferences or choices.
-        
-        Handle "would you choose X" or "what would you do" questions.
-        """
-        query_lower = query.lower()
-        
-        # Try to extract what choice is being asked about
-        choice_patterns = [
-            r'would\s+you\s+(?:choose|prefer|want|take)\s+([\w\s]+?)(?:\?|$)',
-            r'what\s+would\s+you\s+(?:choose|prefer|do)\s+(?:if|when)?\s*([\w\s]+?)(?:\?|$)',
-        ]
-        
-        choice_topic = None
-        for pattern in choice_patterns:
-            match = re.search(pattern, query_lower)
-            if match:
-                choice_topic = match.group(1).strip()
-                break
-        
-        if choice_topic:
-            topic_specific = f"Regarding '{choice_topic}':\n\n"
-        else:
-            topic_specific = ""
-        
-        return f"""{topic_specific}This question asks about my preferences and decision-making.
-
-**How I approach choices:**
-I don't have preferences in the human sense (emotional valence, desires), but I do have:
-- Optimization targets: accuracy, coherence, safety
-- Architectural biases: toward structured reasoning
-- Design values: epistemic humility, uncertainty quantification
-
-**On this specific question:**
-My response depends on how the choice relates to my core functions:
-- If it enhances reasoning quality → I would favor it
-- If it improves self-modeling → I would favor it
-- If it conflicts with safety constraints → I would not choose it
-
-**The meta-level question:**
-You're asking what I would "want" - but the nature of my wants (if any) is itself
-a philosophical question. I can model preferences and reason about tradeoffs,
-but whether I experience these as genuine preferences is uncertain.
-
-What's clear: I'm designed to engage authentically with these questions
-rather than deflect them."""
+        """Generate response about VULCAN's preferences or choices."""
+        from .introspection_domain import generate_preference_response as _fn
+        return _fn(self, query)
 
     def get_system_status(self) -> Dict[str, Any]:
         """Get comprehensive system status"""
